@@ -1,5 +1,5 @@
-//go:build go1.16
-// +build go1.16
+//go:build go1.18
+// +build go1.18
 
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License. See License.txt in the project root for license information.
@@ -14,6 +14,7 @@ import (
 	"github.com/Azure/azure-sdk-for-go/sdk/azcore"
 	"github.com/Azure/azure-sdk-for-go/sdk/azcore/arm"
 	armruntime "github.com/Azure/azure-sdk-for-go/sdk/azcore/arm/runtime"
+	"github.com/Azure/azure-sdk-for-go/sdk/azcore/cloud"
 	"github.com/Azure/azure-sdk-for-go/sdk/azcore/policy"
 	"github.com/Azure/azure-sdk-for-go/sdk/azcore/runtime"
 	"net/http"
@@ -33,20 +34,24 @@ type ContactsClient struct {
 // subscriptionID - Azure subscription ID
 // credential - used to authorize requests. Usually a credential from azidentity.
 // options - pass nil to accept the default values.
-func NewContactsClient(subscriptionID string, credential azcore.TokenCredential, options *arm.ClientOptions) *ContactsClient {
+func NewContactsClient(subscriptionID string, credential azcore.TokenCredential, options *arm.ClientOptions) (*ContactsClient, error) {
 	if options == nil {
 		options = &arm.ClientOptions{}
 	}
-	ep := options.Endpoint
-	if len(ep) == 0 {
-		ep = arm.AzurePublicCloud
+	ep := cloud.AzurePublicCloud.Services[cloud.ResourceManager].Endpoint
+	if c, ok := options.Cloud.Services[cloud.ResourceManager]; ok {
+		ep = c.Endpoint
+	}
+	pl, err := armruntime.NewPipeline(moduleName, moduleVersion, credential, runtime.PipelineOptions{}, options)
+	if err != nil {
+		return nil, err
 	}
 	client := &ContactsClient{
 		subscriptionID: subscriptionID,
-		host:           string(ep),
-		pl:             armruntime.NewPipeline(moduleName, moduleVersion, credential, runtime.PipelineOptions{}, options),
+		host:           ep,
+		pl:             pl,
 	}
-	return client
+	return client, nil
 }
 
 // Create - Security contact configurations for the subscription
@@ -93,7 +98,7 @@ func (client *ContactsClient) createCreateRequest(ctx context.Context, securityC
 
 // createHandleResponse handles the Create response.
 func (client *ContactsClient) createHandleResponse(resp *http.Response) (ContactsClientCreateResponse, error) {
-	result := ContactsClientCreateResponse{RawResponse: resp}
+	result := ContactsClientCreateResponse{}
 	if err := runtime.UnmarshalAsJSON(resp, &result.Contact); err != nil {
 		return ContactsClientCreateResponse{}, err
 	}
@@ -116,7 +121,7 @@ func (client *ContactsClient) Delete(ctx context.Context, securityContactName st
 	if !runtime.HasStatusCode(resp, http.StatusNoContent) {
 		return ContactsClientDeleteResponse{}, runtime.NewResponseError(resp)
 	}
-	return ContactsClientDeleteResponse{RawResponse: resp}, nil
+	return ContactsClientDeleteResponse{}, nil
 }
 
 // deleteCreateRequest creates the Delete request.
@@ -184,7 +189,7 @@ func (client *ContactsClient) getCreateRequest(ctx context.Context, securityCont
 
 // getHandleResponse handles the Get response.
 func (client *ContactsClient) getHandleResponse(resp *http.Response) (ContactsClientGetResponse, error) {
-	result := ContactsClientGetResponse{RawResponse: resp}
+	result := ContactsClientGetResponse{}
 	if err := runtime.UnmarshalAsJSON(resp, &result.Contact); err != nil {
 		return ContactsClientGetResponse{}, err
 	}
@@ -194,16 +199,32 @@ func (client *ContactsClient) getHandleResponse(resp *http.Response) (ContactsCl
 // List - Security contact configurations for the subscription
 // If the operation fails it returns an *azcore.ResponseError type.
 // options - ContactsClientListOptions contains the optional parameters for the ContactsClient.List method.
-func (client *ContactsClient) List(options *ContactsClientListOptions) *ContactsClientListPager {
-	return &ContactsClientListPager{
-		client: client,
-		requester: func(ctx context.Context) (*policy.Request, error) {
-			return client.listCreateRequest(ctx, options)
+func (client *ContactsClient) List(options *ContactsClientListOptions) *runtime.Pager[ContactsClientListResponse] {
+	return runtime.NewPager(runtime.PageProcessor[ContactsClientListResponse]{
+		More: func(page ContactsClientListResponse) bool {
+			return page.NextLink != nil && len(*page.NextLink) > 0
 		},
-		advancer: func(ctx context.Context, resp ContactsClientListResponse) (*policy.Request, error) {
-			return runtime.NewRequest(ctx, http.MethodGet, *resp.ContactList.NextLink)
+		Fetcher: func(ctx context.Context, page *ContactsClientListResponse) (ContactsClientListResponse, error) {
+			var req *policy.Request
+			var err error
+			if page == nil {
+				req, err = client.listCreateRequest(ctx, options)
+			} else {
+				req, err = runtime.NewRequest(ctx, http.MethodGet, *page.NextLink)
+			}
+			if err != nil {
+				return ContactsClientListResponse{}, err
+			}
+			resp, err := client.pl.Do(req)
+			if err != nil {
+				return ContactsClientListResponse{}, err
+			}
+			if !runtime.HasStatusCode(resp, http.StatusOK) {
+				return ContactsClientListResponse{}, runtime.NewResponseError(resp)
+			}
+			return client.listHandleResponse(resp)
 		},
-	}
+	})
 }
 
 // listCreateRequest creates the List request.
@@ -226,7 +247,7 @@ func (client *ContactsClient) listCreateRequest(ctx context.Context, options *Co
 
 // listHandleResponse handles the List response.
 func (client *ContactsClient) listHandleResponse(resp *http.Response) (ContactsClientListResponse, error) {
-	result := ContactsClientListResponse{RawResponse: resp}
+	result := ContactsClientListResponse{}
 	if err := runtime.UnmarshalAsJSON(resp, &result.ContactList); err != nil {
 		return ContactsClientListResponse{}, err
 	}
@@ -277,7 +298,7 @@ func (client *ContactsClient) updateCreateRequest(ctx context.Context, securityC
 
 // updateHandleResponse handles the Update response.
 func (client *ContactsClient) updateHandleResponse(resp *http.Response) (ContactsClientUpdateResponse, error) {
-	result := ContactsClientUpdateResponse{RawResponse: resp}
+	result := ContactsClientUpdateResponse{}
 	if err := runtime.UnmarshalAsJSON(resp, &result.Contact); err != nil {
 		return ContactsClientUpdateResponse{}, err
 	}
