@@ -1,5 +1,5 @@
-//go:build go1.16
-// +build go1.16
+//go:build go1.18
+// +build go1.18
 
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License. See License.txt in the project root for license information.
@@ -14,6 +14,7 @@ import (
 	"github.com/Azure/azure-sdk-for-go/sdk/azcore"
 	"github.com/Azure/azure-sdk-for-go/sdk/azcore/arm"
 	armruntime "github.com/Azure/azure-sdk-for-go/sdk/azcore/arm/runtime"
+	"github.com/Azure/azure-sdk-for-go/sdk/azcore/cloud"
 	"github.com/Azure/azure-sdk-for-go/sdk/azcore/policy"
 	"github.com/Azure/azure-sdk-for-go/sdk/azcore/runtime"
 	"net/http"
@@ -34,20 +35,24 @@ type IotSecuritySolutionsAnalyticsRecommendationClient struct {
 // subscriptionID - Azure subscription ID
 // credential - used to authorize requests. Usually a credential from azidentity.
 // options - pass nil to accept the default values.
-func NewIotSecuritySolutionsAnalyticsRecommendationClient(subscriptionID string, credential azcore.TokenCredential, options *arm.ClientOptions) *IotSecuritySolutionsAnalyticsRecommendationClient {
+func NewIotSecuritySolutionsAnalyticsRecommendationClient(subscriptionID string, credential azcore.TokenCredential, options *arm.ClientOptions) (*IotSecuritySolutionsAnalyticsRecommendationClient, error) {
 	if options == nil {
 		options = &arm.ClientOptions{}
 	}
-	ep := options.Endpoint
-	if len(ep) == 0 {
-		ep = arm.AzurePublicCloud
+	ep := cloud.AzurePublicCloud.Services[cloud.ResourceManager].Endpoint
+	if c, ok := options.Cloud.Services[cloud.ResourceManager]; ok {
+		ep = c.Endpoint
+	}
+	pl, err := armruntime.NewPipeline(moduleName, moduleVersion, credential, runtime.PipelineOptions{}, options)
+	if err != nil {
+		return nil, err
 	}
 	client := &IotSecuritySolutionsAnalyticsRecommendationClient{
 		subscriptionID: subscriptionID,
-		host:           string(ep),
-		pl:             armruntime.NewPipeline(moduleName, moduleVersion, credential, runtime.PipelineOptions{}, options),
+		host:           ep,
+		pl:             pl,
 	}
-	return client
+	return client, nil
 }
 
 // Get - Use this method to get the aggregated security analytics recommendation of yours IoT Security solution. This aggregation
@@ -105,7 +110,7 @@ func (client *IotSecuritySolutionsAnalyticsRecommendationClient) getCreateReques
 
 // getHandleResponse handles the Get response.
 func (client *IotSecuritySolutionsAnalyticsRecommendationClient) getHandleResponse(resp *http.Response) (IotSecuritySolutionsAnalyticsRecommendationClientGetResponse, error) {
-	result := IotSecuritySolutionsAnalyticsRecommendationClientGetResponse{RawResponse: resp}
+	result := IotSecuritySolutionsAnalyticsRecommendationClientGetResponse{}
 	if err := runtime.UnmarshalAsJSON(resp, &result.IoTSecurityAggregatedRecommendation); err != nil {
 		return IotSecuritySolutionsAnalyticsRecommendationClientGetResponse{}, err
 	}
@@ -118,16 +123,32 @@ func (client *IotSecuritySolutionsAnalyticsRecommendationClient) getHandleRespon
 // solutionName - The name of the IoT Security solution.
 // options - IotSecuritySolutionsAnalyticsRecommendationClientListOptions contains the optional parameters for the IotSecuritySolutionsAnalyticsRecommendationClient.List
 // method.
-func (client *IotSecuritySolutionsAnalyticsRecommendationClient) List(resourceGroupName string, solutionName string, options *IotSecuritySolutionsAnalyticsRecommendationClientListOptions) *IotSecuritySolutionsAnalyticsRecommendationClientListPager {
-	return &IotSecuritySolutionsAnalyticsRecommendationClientListPager{
-		client: client,
-		requester: func(ctx context.Context) (*policy.Request, error) {
-			return client.listCreateRequest(ctx, resourceGroupName, solutionName, options)
+func (client *IotSecuritySolutionsAnalyticsRecommendationClient) List(resourceGroupName string, solutionName string, options *IotSecuritySolutionsAnalyticsRecommendationClientListOptions) *runtime.Pager[IotSecuritySolutionsAnalyticsRecommendationClientListResponse] {
+	return runtime.NewPager(runtime.PageProcessor[IotSecuritySolutionsAnalyticsRecommendationClientListResponse]{
+		More: func(page IotSecuritySolutionsAnalyticsRecommendationClientListResponse) bool {
+			return page.NextLink != nil && len(*page.NextLink) > 0
 		},
-		advancer: func(ctx context.Context, resp IotSecuritySolutionsAnalyticsRecommendationClientListResponse) (*policy.Request, error) {
-			return runtime.NewRequest(ctx, http.MethodGet, *resp.IoTSecurityAggregatedRecommendationList.NextLink)
+		Fetcher: func(ctx context.Context, page *IotSecuritySolutionsAnalyticsRecommendationClientListResponse) (IotSecuritySolutionsAnalyticsRecommendationClientListResponse, error) {
+			var req *policy.Request
+			var err error
+			if page == nil {
+				req, err = client.listCreateRequest(ctx, resourceGroupName, solutionName, options)
+			} else {
+				req, err = runtime.NewRequest(ctx, http.MethodGet, *page.NextLink)
+			}
+			if err != nil {
+				return IotSecuritySolutionsAnalyticsRecommendationClientListResponse{}, err
+			}
+			resp, err := client.pl.Do(req)
+			if err != nil {
+				return IotSecuritySolutionsAnalyticsRecommendationClientListResponse{}, err
+			}
+			if !runtime.HasStatusCode(resp, http.StatusOK) {
+				return IotSecuritySolutionsAnalyticsRecommendationClientListResponse{}, runtime.NewResponseError(resp)
+			}
+			return client.listHandleResponse(resp)
 		},
-	}
+	})
 }
 
 // listCreateRequest creates the List request.
@@ -161,7 +182,7 @@ func (client *IotSecuritySolutionsAnalyticsRecommendationClient) listCreateReque
 
 // listHandleResponse handles the List response.
 func (client *IotSecuritySolutionsAnalyticsRecommendationClient) listHandleResponse(resp *http.Response) (IotSecuritySolutionsAnalyticsRecommendationClientListResponse, error) {
-	result := IotSecuritySolutionsAnalyticsRecommendationClientListResponse{RawResponse: resp}
+	result := IotSecuritySolutionsAnalyticsRecommendationClientListResponse{}
 	if err := runtime.UnmarshalAsJSON(resp, &result.IoTSecurityAggregatedRecommendationList); err != nil {
 		return IotSecuritySolutionsAnalyticsRecommendationClientListResponse{}, err
 	}
