@@ -1,5 +1,5 @@
-//go:build go1.16
-// +build go1.16
+//go:build go1.18
+// +build go1.18
 
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License. See License.txt in the project root for license information.
@@ -8,1797 +8,763 @@
 
 package armsecurity
 
-import (
-	"context"
-	armruntime "github.com/Azure/azure-sdk-for-go/sdk/azcore/arm/runtime"
-	"net/http"
-	"time"
-)
-
 // AccountConnectorsClientCreateOrUpdateResponse contains the response from method AccountConnectorsClient.CreateOrUpdate.
 type AccountConnectorsClientCreateOrUpdateResponse struct {
-	AccountConnectorsClientCreateOrUpdateResult
-	// RawResponse contains the underlying HTTP response.
-	RawResponse *http.Response
-}
-
-// AccountConnectorsClientCreateOrUpdateResult contains the result from method AccountConnectorsClient.CreateOrUpdate.
-type AccountConnectorsClientCreateOrUpdateResult struct {
 	ConnectorSetting
 }
 
 // AccountConnectorsClientDeleteResponse contains the response from method AccountConnectorsClient.Delete.
 type AccountConnectorsClientDeleteResponse struct {
-	// RawResponse contains the underlying HTTP response.
-	RawResponse *http.Response
+	// placeholder for future response values
 }
 
 // AccountConnectorsClientGetResponse contains the response from method AccountConnectorsClient.Get.
 type AccountConnectorsClientGetResponse struct {
-	AccountConnectorsClientGetResult
-	// RawResponse contains the underlying HTTP response.
-	RawResponse *http.Response
-}
-
-// AccountConnectorsClientGetResult contains the result from method AccountConnectorsClient.Get.
-type AccountConnectorsClientGetResult struct {
 	ConnectorSetting
 }
 
 // AccountConnectorsClientListResponse contains the response from method AccountConnectorsClient.List.
 type AccountConnectorsClientListResponse struct {
-	AccountConnectorsClientListResult
-	// RawResponse contains the underlying HTTP response.
-	RawResponse *http.Response
-}
-
-// AccountConnectorsClientListResult contains the result from method AccountConnectorsClient.List.
-type AccountConnectorsClientListResult struct {
 	ConnectorSettingList
 }
 
 // AdaptiveApplicationControlsClientDeleteResponse contains the response from method AdaptiveApplicationControlsClient.Delete.
 type AdaptiveApplicationControlsClientDeleteResponse struct {
-	// RawResponse contains the underlying HTTP response.
-	RawResponse *http.Response
+	// placeholder for future response values
 }
 
 // AdaptiveApplicationControlsClientGetResponse contains the response from method AdaptiveApplicationControlsClient.Get.
 type AdaptiveApplicationControlsClientGetResponse struct {
-	AdaptiveApplicationControlsClientGetResult
-	// RawResponse contains the underlying HTTP response.
-	RawResponse *http.Response
-}
-
-// AdaptiveApplicationControlsClientGetResult contains the result from method AdaptiveApplicationControlsClient.Get.
-type AdaptiveApplicationControlsClientGetResult struct {
 	AdaptiveApplicationControlGroup
 }
 
 // AdaptiveApplicationControlsClientListResponse contains the response from method AdaptiveApplicationControlsClient.List.
 type AdaptiveApplicationControlsClientListResponse struct {
-	AdaptiveApplicationControlsClientListResult
-	// RawResponse contains the underlying HTTP response.
-	RawResponse *http.Response
-}
-
-// AdaptiveApplicationControlsClientListResult contains the result from method AdaptiveApplicationControlsClient.List.
-type AdaptiveApplicationControlsClientListResult struct {
 	AdaptiveApplicationControlGroups
 }
 
 // AdaptiveApplicationControlsClientPutResponse contains the response from method AdaptiveApplicationControlsClient.Put.
 type AdaptiveApplicationControlsClientPutResponse struct {
-	AdaptiveApplicationControlsClientPutResult
-	// RawResponse contains the underlying HTTP response.
-	RawResponse *http.Response
-}
-
-// AdaptiveApplicationControlsClientPutResult contains the result from method AdaptiveApplicationControlsClient.Put.
-type AdaptiveApplicationControlsClientPutResult struct {
 	AdaptiveApplicationControlGroup
-}
-
-// AdaptiveNetworkHardeningsClientEnforcePollerResponse contains the response from method AdaptiveNetworkHardeningsClient.Enforce.
-type AdaptiveNetworkHardeningsClientEnforcePollerResponse struct {
-	// Poller contains an initialized poller.
-	Poller *AdaptiveNetworkHardeningsClientEnforcePoller
-
-	// RawResponse contains the underlying HTTP response.
-	RawResponse *http.Response
-}
-
-// PollUntilDone will poll the service endpoint until a terminal state is reached or an error is received.
-// freq: the time to wait between intervals in absence of a Retry-After header. Allowed minimum is one second.
-// A good starting value is 30 seconds. Note that some resources might benefit from a different value.
-func (l AdaptiveNetworkHardeningsClientEnforcePollerResponse) PollUntilDone(ctx context.Context, freq time.Duration) (AdaptiveNetworkHardeningsClientEnforceResponse, error) {
-	respType := AdaptiveNetworkHardeningsClientEnforceResponse{}
-	resp, err := l.Poller.pt.PollUntilDone(ctx, freq, nil)
-	if err != nil {
-		return respType, err
-	}
-	respType.RawResponse = resp
-	return respType, nil
-}
-
-// Resume rehydrates a AdaptiveNetworkHardeningsClientEnforcePollerResponse from the provided client and resume token.
-func (l *AdaptiveNetworkHardeningsClientEnforcePollerResponse) Resume(ctx context.Context, client *AdaptiveNetworkHardeningsClient, token string) error {
-	pt, err := armruntime.NewPollerFromResumeToken("AdaptiveNetworkHardeningsClient.Enforce", token, client.pl)
-	if err != nil {
-		return err
-	}
-	poller := &AdaptiveNetworkHardeningsClientEnforcePoller{
-		pt: pt,
-	}
-	resp, err := poller.Poll(ctx)
-	if err != nil {
-		return err
-	}
-	l.Poller = poller
-	l.RawResponse = resp
-	return nil
 }
 
 // AdaptiveNetworkHardeningsClientEnforceResponse contains the response from method AdaptiveNetworkHardeningsClient.Enforce.
 type AdaptiveNetworkHardeningsClientEnforceResponse struct {
-	// RawResponse contains the underlying HTTP response.
-	RawResponse *http.Response
+	// placeholder for future response values
 }
 
 // AdaptiveNetworkHardeningsClientGetResponse contains the response from method AdaptiveNetworkHardeningsClient.Get.
 type AdaptiveNetworkHardeningsClientGetResponse struct {
-	AdaptiveNetworkHardeningsClientGetResult
-	// RawResponse contains the underlying HTTP response.
-	RawResponse *http.Response
-}
-
-// AdaptiveNetworkHardeningsClientGetResult contains the result from method AdaptiveNetworkHardeningsClient.Get.
-type AdaptiveNetworkHardeningsClientGetResult struct {
 	AdaptiveNetworkHardening
 }
 
 // AdaptiveNetworkHardeningsClientListByExtendedResourceResponse contains the response from method AdaptiveNetworkHardeningsClient.ListByExtendedResource.
 type AdaptiveNetworkHardeningsClientListByExtendedResourceResponse struct {
-	AdaptiveNetworkHardeningsClientListByExtendedResourceResult
-	// RawResponse contains the underlying HTTP response.
-	RawResponse *http.Response
-}
-
-// AdaptiveNetworkHardeningsClientListByExtendedResourceResult contains the result from method AdaptiveNetworkHardeningsClient.ListByExtendedResource.
-type AdaptiveNetworkHardeningsClientListByExtendedResourceResult struct {
 	AdaptiveNetworkHardeningsList
 }
 
 // AdvancedThreatProtectionClientCreateResponse contains the response from method AdvancedThreatProtectionClient.Create.
 type AdvancedThreatProtectionClientCreateResponse struct {
-	AdvancedThreatProtectionClientCreateResult
-	// RawResponse contains the underlying HTTP response.
-	RawResponse *http.Response
-}
-
-// AdvancedThreatProtectionClientCreateResult contains the result from method AdvancedThreatProtectionClient.Create.
-type AdvancedThreatProtectionClientCreateResult struct {
 	AdvancedThreatProtectionSetting
 }
 
 // AdvancedThreatProtectionClientGetResponse contains the response from method AdvancedThreatProtectionClient.Get.
 type AdvancedThreatProtectionClientGetResponse struct {
-	AdvancedThreatProtectionClientGetResult
-	// RawResponse contains the underlying HTTP response.
-	RawResponse *http.Response
-}
-
-// AdvancedThreatProtectionClientGetResult contains the result from method AdvancedThreatProtectionClient.Get.
-type AdvancedThreatProtectionClientGetResult struct {
 	AdvancedThreatProtectionSetting
 }
 
 // AlertsClientGetResourceGroupLevelResponse contains the response from method AlertsClient.GetResourceGroupLevel.
 type AlertsClientGetResourceGroupLevelResponse struct {
-	AlertsClientGetResourceGroupLevelResult
-	// RawResponse contains the underlying HTTP response.
-	RawResponse *http.Response
-}
-
-// AlertsClientGetResourceGroupLevelResult contains the result from method AlertsClient.GetResourceGroupLevel.
-type AlertsClientGetResourceGroupLevelResult struct {
 	Alert
 }
 
 // AlertsClientGetSubscriptionLevelResponse contains the response from method AlertsClient.GetSubscriptionLevel.
 type AlertsClientGetSubscriptionLevelResponse struct {
-	AlertsClientGetSubscriptionLevelResult
-	// RawResponse contains the underlying HTTP response.
-	RawResponse *http.Response
-}
-
-// AlertsClientGetSubscriptionLevelResult contains the result from method AlertsClient.GetSubscriptionLevel.
-type AlertsClientGetSubscriptionLevelResult struct {
 	Alert
 }
 
 // AlertsClientListByResourceGroupResponse contains the response from method AlertsClient.ListByResourceGroup.
 type AlertsClientListByResourceGroupResponse struct {
-	AlertsClientListByResourceGroupResult
-	// RawResponse contains the underlying HTTP response.
-	RawResponse *http.Response
-}
-
-// AlertsClientListByResourceGroupResult contains the result from method AlertsClient.ListByResourceGroup.
-type AlertsClientListByResourceGroupResult struct {
 	AlertList
 }
 
 // AlertsClientListResourceGroupLevelByRegionResponse contains the response from method AlertsClient.ListResourceGroupLevelByRegion.
 type AlertsClientListResourceGroupLevelByRegionResponse struct {
-	AlertsClientListResourceGroupLevelByRegionResult
-	// RawResponse contains the underlying HTTP response.
-	RawResponse *http.Response
-}
-
-// AlertsClientListResourceGroupLevelByRegionResult contains the result from method AlertsClient.ListResourceGroupLevelByRegion.
-type AlertsClientListResourceGroupLevelByRegionResult struct {
 	AlertList
 }
 
 // AlertsClientListResponse contains the response from method AlertsClient.List.
 type AlertsClientListResponse struct {
-	AlertsClientListResult
-	// RawResponse contains the underlying HTTP response.
-	RawResponse *http.Response
-}
-
-// AlertsClientListResult contains the result from method AlertsClient.List.
-type AlertsClientListResult struct {
 	AlertList
 }
 
 // AlertsClientListSubscriptionLevelByRegionResponse contains the response from method AlertsClient.ListSubscriptionLevelByRegion.
 type AlertsClientListSubscriptionLevelByRegionResponse struct {
-	AlertsClientListSubscriptionLevelByRegionResult
-	// RawResponse contains the underlying HTTP response.
-	RawResponse *http.Response
-}
-
-// AlertsClientListSubscriptionLevelByRegionResult contains the result from method AlertsClient.ListSubscriptionLevelByRegion.
-type AlertsClientListSubscriptionLevelByRegionResult struct {
 	AlertList
-}
-
-// AlertsClientSimulatePollerResponse contains the response from method AlertsClient.Simulate.
-type AlertsClientSimulatePollerResponse struct {
-	// Poller contains an initialized poller.
-	Poller *AlertsClientSimulatePoller
-
-	// RawResponse contains the underlying HTTP response.
-	RawResponse *http.Response
-}
-
-// PollUntilDone will poll the service endpoint until a terminal state is reached or an error is received.
-// freq: the time to wait between intervals in absence of a Retry-After header. Allowed minimum is one second.
-// A good starting value is 30 seconds. Note that some resources might benefit from a different value.
-func (l AlertsClientSimulatePollerResponse) PollUntilDone(ctx context.Context, freq time.Duration) (AlertsClientSimulateResponse, error) {
-	respType := AlertsClientSimulateResponse{}
-	resp, err := l.Poller.pt.PollUntilDone(ctx, freq, nil)
-	if err != nil {
-		return respType, err
-	}
-	respType.RawResponse = resp
-	return respType, nil
-}
-
-// Resume rehydrates a AlertsClientSimulatePollerResponse from the provided client and resume token.
-func (l *AlertsClientSimulatePollerResponse) Resume(ctx context.Context, client *AlertsClient, token string) error {
-	pt, err := armruntime.NewPollerFromResumeToken("AlertsClient.Simulate", token, client.pl)
-	if err != nil {
-		return err
-	}
-	poller := &AlertsClientSimulatePoller{
-		pt: pt,
-	}
-	resp, err := poller.Poll(ctx)
-	if err != nil {
-		return err
-	}
-	l.Poller = poller
-	l.RawResponse = resp
-	return nil
 }
 
 // AlertsClientSimulateResponse contains the response from method AlertsClient.Simulate.
 type AlertsClientSimulateResponse struct {
-	// RawResponse contains the underlying HTTP response.
-	RawResponse *http.Response
+	// placeholder for future response values
 }
 
 // AlertsClientUpdateResourceGroupLevelStateToActivateResponse contains the response from method AlertsClient.UpdateResourceGroupLevelStateToActivate.
 type AlertsClientUpdateResourceGroupLevelStateToActivateResponse struct {
-	// RawResponse contains the underlying HTTP response.
-	RawResponse *http.Response
+	// placeholder for future response values
 }
 
 // AlertsClientUpdateResourceGroupLevelStateToDismissResponse contains the response from method AlertsClient.UpdateResourceGroupLevelStateToDismiss.
 type AlertsClientUpdateResourceGroupLevelStateToDismissResponse struct {
-	// RawResponse contains the underlying HTTP response.
-	RawResponse *http.Response
+	// placeholder for future response values
 }
 
 // AlertsClientUpdateResourceGroupLevelStateToResolveResponse contains the response from method AlertsClient.UpdateResourceGroupLevelStateToResolve.
 type AlertsClientUpdateResourceGroupLevelStateToResolveResponse struct {
-	// RawResponse contains the underlying HTTP response.
-	RawResponse *http.Response
+	// placeholder for future response values
 }
 
 // AlertsClientUpdateSubscriptionLevelStateToActivateResponse contains the response from method AlertsClient.UpdateSubscriptionLevelStateToActivate.
 type AlertsClientUpdateSubscriptionLevelStateToActivateResponse struct {
-	// RawResponse contains the underlying HTTP response.
-	RawResponse *http.Response
+	// placeholder for future response values
 }
 
 // AlertsClientUpdateSubscriptionLevelStateToDismissResponse contains the response from method AlertsClient.UpdateSubscriptionLevelStateToDismiss.
 type AlertsClientUpdateSubscriptionLevelStateToDismissResponse struct {
-	// RawResponse contains the underlying HTTP response.
-	RawResponse *http.Response
+	// placeholder for future response values
 }
 
 // AlertsClientUpdateSubscriptionLevelStateToResolveResponse contains the response from method AlertsClient.UpdateSubscriptionLevelStateToResolve.
 type AlertsClientUpdateSubscriptionLevelStateToResolveResponse struct {
-	// RawResponse contains the underlying HTTP response.
-	RawResponse *http.Response
+	// placeholder for future response values
 }
 
 // AlertsSuppressionRulesClientDeleteResponse contains the response from method AlertsSuppressionRulesClient.Delete.
 type AlertsSuppressionRulesClientDeleteResponse struct {
-	// RawResponse contains the underlying HTTP response.
-	RawResponse *http.Response
+	// placeholder for future response values
 }
 
 // AlertsSuppressionRulesClientGetResponse contains the response from method AlertsSuppressionRulesClient.Get.
 type AlertsSuppressionRulesClientGetResponse struct {
-	AlertsSuppressionRulesClientGetResult
-	// RawResponse contains the underlying HTTP response.
-	RawResponse *http.Response
-}
-
-// AlertsSuppressionRulesClientGetResult contains the result from method AlertsSuppressionRulesClient.Get.
-type AlertsSuppressionRulesClientGetResult struct {
 	AlertsSuppressionRule
 }
 
 // AlertsSuppressionRulesClientListResponse contains the response from method AlertsSuppressionRulesClient.List.
 type AlertsSuppressionRulesClientListResponse struct {
-	AlertsSuppressionRulesClientListResult
-	// RawResponse contains the underlying HTTP response.
-	RawResponse *http.Response
-}
-
-// AlertsSuppressionRulesClientListResult contains the result from method AlertsSuppressionRulesClient.List.
-type AlertsSuppressionRulesClientListResult struct {
 	AlertsSuppressionRulesList
 }
 
 // AlertsSuppressionRulesClientUpdateResponse contains the response from method AlertsSuppressionRulesClient.Update.
 type AlertsSuppressionRulesClientUpdateResponse struct {
-	AlertsSuppressionRulesClientUpdateResult
-	// RawResponse contains the underlying HTTP response.
-	RawResponse *http.Response
-}
-
-// AlertsSuppressionRulesClientUpdateResult contains the result from method AlertsSuppressionRulesClient.Update.
-type AlertsSuppressionRulesClientUpdateResult struct {
 	AlertsSuppressionRule
 }
 
 // AllowedConnectionsClientGetResponse contains the response from method AllowedConnectionsClient.Get.
 type AllowedConnectionsClientGetResponse struct {
-	AllowedConnectionsClientGetResult
-	// RawResponse contains the underlying HTTP response.
-	RawResponse *http.Response
-}
-
-// AllowedConnectionsClientGetResult contains the result from method AllowedConnectionsClient.Get.
-type AllowedConnectionsClientGetResult struct {
 	AllowedConnectionsResource
 }
 
 // AllowedConnectionsClientListByHomeRegionResponse contains the response from method AllowedConnectionsClient.ListByHomeRegion.
 type AllowedConnectionsClientListByHomeRegionResponse struct {
-	AllowedConnectionsClientListByHomeRegionResult
-	// RawResponse contains the underlying HTTP response.
-	RawResponse *http.Response
-}
-
-// AllowedConnectionsClientListByHomeRegionResult contains the result from method AllowedConnectionsClient.ListByHomeRegion.
-type AllowedConnectionsClientListByHomeRegionResult struct {
 	AllowedConnectionsList
 }
 
 // AllowedConnectionsClientListResponse contains the response from method AllowedConnectionsClient.List.
 type AllowedConnectionsClientListResponse struct {
-	AllowedConnectionsClientListResult
-	// RawResponse contains the underlying HTTP response.
-	RawResponse *http.Response
-}
-
-// AllowedConnectionsClientListResult contains the result from method AllowedConnectionsClient.List.
-type AllowedConnectionsClientListResult struct {
 	AllowedConnectionsList
 }
 
 // AssessmentsClientCreateOrUpdateResponse contains the response from method AssessmentsClient.CreateOrUpdate.
 type AssessmentsClientCreateOrUpdateResponse struct {
-	AssessmentsClientCreateOrUpdateResult
-	// RawResponse contains the underlying HTTP response.
-	RawResponse *http.Response
-}
-
-// AssessmentsClientCreateOrUpdateResult contains the result from method AssessmentsClient.CreateOrUpdate.
-type AssessmentsClientCreateOrUpdateResult struct {
 	AssessmentResponse
 }
 
 // AssessmentsClientDeleteResponse contains the response from method AssessmentsClient.Delete.
 type AssessmentsClientDeleteResponse struct {
-	// RawResponse contains the underlying HTTP response.
-	RawResponse *http.Response
+	// placeholder for future response values
 }
 
 // AssessmentsClientGetResponse contains the response from method AssessmentsClient.Get.
 type AssessmentsClientGetResponse struct {
-	AssessmentsClientGetResult
-	// RawResponse contains the underlying HTTP response.
-	RawResponse *http.Response
-}
-
-// AssessmentsClientGetResult contains the result from method AssessmentsClient.Get.
-type AssessmentsClientGetResult struct {
 	AssessmentResponse
 }
 
 // AssessmentsClientListResponse contains the response from method AssessmentsClient.List.
 type AssessmentsClientListResponse struct {
-	AssessmentsClientListResult
-	// RawResponse contains the underlying HTTP response.
-	RawResponse *http.Response
-}
-
-// AssessmentsClientListResult contains the result from method AssessmentsClient.List.
-type AssessmentsClientListResult struct {
 	AssessmentList
 }
 
 // AssessmentsMetadataClientCreateInSubscriptionResponse contains the response from method AssessmentsMetadataClient.CreateInSubscription.
 type AssessmentsMetadataClientCreateInSubscriptionResponse struct {
-	AssessmentsMetadataClientCreateInSubscriptionResult
-	// RawResponse contains the underlying HTTP response.
-	RawResponse *http.Response
-}
-
-// AssessmentsMetadataClientCreateInSubscriptionResult contains the result from method AssessmentsMetadataClient.CreateInSubscription.
-type AssessmentsMetadataClientCreateInSubscriptionResult struct {
 	AssessmentMetadataResponse
 }
 
 // AssessmentsMetadataClientDeleteInSubscriptionResponse contains the response from method AssessmentsMetadataClient.DeleteInSubscription.
 type AssessmentsMetadataClientDeleteInSubscriptionResponse struct {
-	// RawResponse contains the underlying HTTP response.
-	RawResponse *http.Response
+	// placeholder for future response values
 }
 
 // AssessmentsMetadataClientGetInSubscriptionResponse contains the response from method AssessmentsMetadataClient.GetInSubscription.
 type AssessmentsMetadataClientGetInSubscriptionResponse struct {
-	AssessmentsMetadataClientGetInSubscriptionResult
-	// RawResponse contains the underlying HTTP response.
-	RawResponse *http.Response
-}
-
-// AssessmentsMetadataClientGetInSubscriptionResult contains the result from method AssessmentsMetadataClient.GetInSubscription.
-type AssessmentsMetadataClientGetInSubscriptionResult struct {
 	AssessmentMetadataResponse
 }
 
 // AssessmentsMetadataClientGetResponse contains the response from method AssessmentsMetadataClient.Get.
 type AssessmentsMetadataClientGetResponse struct {
-	AssessmentsMetadataClientGetResult
-	// RawResponse contains the underlying HTTP response.
-	RawResponse *http.Response
-}
-
-// AssessmentsMetadataClientGetResult contains the result from method AssessmentsMetadataClient.Get.
-type AssessmentsMetadataClientGetResult struct {
 	AssessmentMetadataResponse
 }
 
 // AssessmentsMetadataClientListBySubscriptionResponse contains the response from method AssessmentsMetadataClient.ListBySubscription.
 type AssessmentsMetadataClientListBySubscriptionResponse struct {
-	AssessmentsMetadataClientListBySubscriptionResult
-	// RawResponse contains the underlying HTTP response.
-	RawResponse *http.Response
-}
-
-// AssessmentsMetadataClientListBySubscriptionResult contains the result from method AssessmentsMetadataClient.ListBySubscription.
-type AssessmentsMetadataClientListBySubscriptionResult struct {
 	AssessmentMetadataResponseList
 }
 
 // AssessmentsMetadataClientListResponse contains the response from method AssessmentsMetadataClient.List.
 type AssessmentsMetadataClientListResponse struct {
-	AssessmentsMetadataClientListResult
-	// RawResponse contains the underlying HTTP response.
-	RawResponse *http.Response
-}
-
-// AssessmentsMetadataClientListResult contains the result from method AssessmentsMetadataClient.List.
-type AssessmentsMetadataClientListResult struct {
 	AssessmentMetadataResponseList
 }
 
 // AutoProvisioningSettingsClientCreateResponse contains the response from method AutoProvisioningSettingsClient.Create.
 type AutoProvisioningSettingsClientCreateResponse struct {
-	AutoProvisioningSettingsClientCreateResult
-	// RawResponse contains the underlying HTTP response.
-	RawResponse *http.Response
-}
-
-// AutoProvisioningSettingsClientCreateResult contains the result from method AutoProvisioningSettingsClient.Create.
-type AutoProvisioningSettingsClientCreateResult struct {
 	AutoProvisioningSetting
 }
 
 // AutoProvisioningSettingsClientGetResponse contains the response from method AutoProvisioningSettingsClient.Get.
 type AutoProvisioningSettingsClientGetResponse struct {
-	AutoProvisioningSettingsClientGetResult
-	// RawResponse contains the underlying HTTP response.
-	RawResponse *http.Response
-}
-
-// AutoProvisioningSettingsClientGetResult contains the result from method AutoProvisioningSettingsClient.Get.
-type AutoProvisioningSettingsClientGetResult struct {
 	AutoProvisioningSetting
 }
 
 // AutoProvisioningSettingsClientListResponse contains the response from method AutoProvisioningSettingsClient.List.
 type AutoProvisioningSettingsClientListResponse struct {
-	AutoProvisioningSettingsClientListResult
-	// RawResponse contains the underlying HTTP response.
-	RawResponse *http.Response
-}
-
-// AutoProvisioningSettingsClientListResult contains the result from method AutoProvisioningSettingsClient.List.
-type AutoProvisioningSettingsClientListResult struct {
 	AutoProvisioningSettingList
 }
 
 // AutomationsClientCreateOrUpdateResponse contains the response from method AutomationsClient.CreateOrUpdate.
 type AutomationsClientCreateOrUpdateResponse struct {
-	AutomationsClientCreateOrUpdateResult
-	// RawResponse contains the underlying HTTP response.
-	RawResponse *http.Response
-}
-
-// AutomationsClientCreateOrUpdateResult contains the result from method AutomationsClient.CreateOrUpdate.
-type AutomationsClientCreateOrUpdateResult struct {
 	Automation
 }
 
 // AutomationsClientDeleteResponse contains the response from method AutomationsClient.Delete.
 type AutomationsClientDeleteResponse struct {
-	// RawResponse contains the underlying HTTP response.
-	RawResponse *http.Response
+	// placeholder for future response values
 }
 
 // AutomationsClientGetResponse contains the response from method AutomationsClient.Get.
 type AutomationsClientGetResponse struct {
-	AutomationsClientGetResult
-	// RawResponse contains the underlying HTTP response.
-	RawResponse *http.Response
-}
-
-// AutomationsClientGetResult contains the result from method AutomationsClient.Get.
-type AutomationsClientGetResult struct {
 	Automation
 }
 
 // AutomationsClientListByResourceGroupResponse contains the response from method AutomationsClient.ListByResourceGroup.
 type AutomationsClientListByResourceGroupResponse struct {
-	AutomationsClientListByResourceGroupResult
-	// RawResponse contains the underlying HTTP response.
-	RawResponse *http.Response
-}
-
-// AutomationsClientListByResourceGroupResult contains the result from method AutomationsClient.ListByResourceGroup.
-type AutomationsClientListByResourceGroupResult struct {
 	AutomationList
 }
 
 // AutomationsClientListResponse contains the response from method AutomationsClient.List.
 type AutomationsClientListResponse struct {
-	AutomationsClientListResult
-	// RawResponse contains the underlying HTTP response.
-	RawResponse *http.Response
-}
-
-// AutomationsClientListResult contains the result from method AutomationsClient.List.
-type AutomationsClientListResult struct {
 	AutomationList
 }
 
 // AutomationsClientValidateResponse contains the response from method AutomationsClient.Validate.
 type AutomationsClientValidateResponse struct {
-	AutomationsClientValidateResult
-	// RawResponse contains the underlying HTTP response.
-	RawResponse *http.Response
-}
-
-// AutomationsClientValidateResult contains the result from method AutomationsClient.Validate.
-type AutomationsClientValidateResult struct {
 	AutomationValidationStatus
 }
 
 // ComplianceResultsClientGetResponse contains the response from method ComplianceResultsClient.Get.
 type ComplianceResultsClientGetResponse struct {
-	ComplianceResultsClientGetResult
-	// RawResponse contains the underlying HTTP response.
-	RawResponse *http.Response
-}
-
-// ComplianceResultsClientGetResult contains the result from method ComplianceResultsClient.Get.
-type ComplianceResultsClientGetResult struct {
 	ComplianceResult
 }
 
 // ComplianceResultsClientListResponse contains the response from method ComplianceResultsClient.List.
 type ComplianceResultsClientListResponse struct {
-	ComplianceResultsClientListResult
-	// RawResponse contains the underlying HTTP response.
-	RawResponse *http.Response
-}
-
-// ComplianceResultsClientListResult contains the result from method ComplianceResultsClient.List.
-type ComplianceResultsClientListResult struct {
 	ComplianceResultList
 }
 
 // CompliancesClientGetResponse contains the response from method CompliancesClient.Get.
 type CompliancesClientGetResponse struct {
-	CompliancesClientGetResult
-	// RawResponse contains the underlying HTTP response.
-	RawResponse *http.Response
-}
-
-// CompliancesClientGetResult contains the result from method CompliancesClient.Get.
-type CompliancesClientGetResult struct {
 	Compliance
 }
 
 // CompliancesClientListResponse contains the response from method CompliancesClient.List.
 type CompliancesClientListResponse struct {
-	CompliancesClientListResult
-	// RawResponse contains the underlying HTTP response.
-	RawResponse *http.Response
-}
-
-// CompliancesClientListResult contains the result from method CompliancesClient.List.
-type CompliancesClientListResult struct {
 	ComplianceList
 }
 
 // ConnectorsClientCreateOrUpdateResponse contains the response from method ConnectorsClient.CreateOrUpdate.
 type ConnectorsClientCreateOrUpdateResponse struct {
-	ConnectorsClientCreateOrUpdateResult
-	// RawResponse contains the underlying HTTP response.
-	RawResponse *http.Response
-}
-
-// ConnectorsClientCreateOrUpdateResult contains the result from method ConnectorsClient.CreateOrUpdate.
-type ConnectorsClientCreateOrUpdateResult struct {
 	Connector
 }
 
 // ConnectorsClientDeleteResponse contains the response from method ConnectorsClient.Delete.
 type ConnectorsClientDeleteResponse struct {
-	// RawResponse contains the underlying HTTP response.
-	RawResponse *http.Response
+	// placeholder for future response values
 }
 
 // ConnectorsClientGetResponse contains the response from method ConnectorsClient.Get.
 type ConnectorsClientGetResponse struct {
-	ConnectorsClientGetResult
-	// RawResponse contains the underlying HTTP response.
-	RawResponse *http.Response
-}
-
-// ConnectorsClientGetResult contains the result from method ConnectorsClient.Get.
-type ConnectorsClientGetResult struct {
 	Connector
 }
 
 // ConnectorsClientListByResourceGroupResponse contains the response from method ConnectorsClient.ListByResourceGroup.
 type ConnectorsClientListByResourceGroupResponse struct {
-	ConnectorsClientListByResourceGroupResult
-	// RawResponse contains the underlying HTTP response.
-	RawResponse *http.Response
-}
-
-// ConnectorsClientListByResourceGroupResult contains the result from method ConnectorsClient.ListByResourceGroup.
-type ConnectorsClientListByResourceGroupResult struct {
 	ConnectorsList
 }
 
 // ConnectorsClientListResponse contains the response from method ConnectorsClient.List.
 type ConnectorsClientListResponse struct {
-	ConnectorsClientListResult
-	// RawResponse contains the underlying HTTP response.
-	RawResponse *http.Response
-}
-
-// ConnectorsClientListResult contains the result from method ConnectorsClient.List.
-type ConnectorsClientListResult struct {
 	ConnectorsList
 }
 
 // ConnectorsClientUpdateResponse contains the response from method ConnectorsClient.Update.
 type ConnectorsClientUpdateResponse struct {
-	ConnectorsClientUpdateResult
-	// RawResponse contains the underlying HTTP response.
-	RawResponse *http.Response
-}
-
-// ConnectorsClientUpdateResult contains the result from method ConnectorsClient.Update.
-type ConnectorsClientUpdateResult struct {
 	Connector
 }
 
 // ContactsClientCreateResponse contains the response from method ContactsClient.Create.
 type ContactsClientCreateResponse struct {
-	ContactsClientCreateResult
-	// RawResponse contains the underlying HTTP response.
-	RawResponse *http.Response
-}
-
-// ContactsClientCreateResult contains the result from method ContactsClient.Create.
-type ContactsClientCreateResult struct {
 	Contact
 }
 
 // ContactsClientDeleteResponse contains the response from method ContactsClient.Delete.
 type ContactsClientDeleteResponse struct {
-	// RawResponse contains the underlying HTTP response.
-	RawResponse *http.Response
+	// placeholder for future response values
 }
 
 // ContactsClientGetResponse contains the response from method ContactsClient.Get.
 type ContactsClientGetResponse struct {
-	ContactsClientGetResult
-	// RawResponse contains the underlying HTTP response.
-	RawResponse *http.Response
-}
-
-// ContactsClientGetResult contains the result from method ContactsClient.Get.
-type ContactsClientGetResult struct {
 	Contact
 }
 
 // ContactsClientListResponse contains the response from method ContactsClient.List.
 type ContactsClientListResponse struct {
-	ContactsClientListResult
-	// RawResponse contains the underlying HTTP response.
-	RawResponse *http.Response
-}
-
-// ContactsClientListResult contains the result from method ContactsClient.List.
-type ContactsClientListResult struct {
 	ContactList
 }
 
 // ContactsClientUpdateResponse contains the response from method ContactsClient.Update.
 type ContactsClientUpdateResponse struct {
-	ContactsClientUpdateResult
-	// RawResponse contains the underlying HTTP response.
-	RawResponse *http.Response
-}
-
-// ContactsClientUpdateResult contains the result from method ContactsClient.Update.
-type ContactsClientUpdateResult struct {
 	Contact
 }
 
 // CustomAssessmentAutomationsClientCreateResponse contains the response from method CustomAssessmentAutomationsClient.Create.
 type CustomAssessmentAutomationsClientCreateResponse struct {
-	CustomAssessmentAutomationsClientCreateResult
-	// RawResponse contains the underlying HTTP response.
-	RawResponse *http.Response
-}
-
-// CustomAssessmentAutomationsClientCreateResult contains the result from method CustomAssessmentAutomationsClient.Create.
-type CustomAssessmentAutomationsClientCreateResult struct {
 	CustomAssessmentAutomation
 }
 
 // CustomAssessmentAutomationsClientDeleteResponse contains the response from method CustomAssessmentAutomationsClient.Delete.
 type CustomAssessmentAutomationsClientDeleteResponse struct {
-	// RawResponse contains the underlying HTTP response.
-	RawResponse *http.Response
+	// placeholder for future response values
 }
 
 // CustomAssessmentAutomationsClientGetResponse contains the response from method CustomAssessmentAutomationsClient.Get.
 type CustomAssessmentAutomationsClientGetResponse struct {
-	CustomAssessmentAutomationsClientGetResult
-	// RawResponse contains the underlying HTTP response.
-	RawResponse *http.Response
-}
-
-// CustomAssessmentAutomationsClientGetResult contains the result from method CustomAssessmentAutomationsClient.Get.
-type CustomAssessmentAutomationsClientGetResult struct {
 	CustomAssessmentAutomation
 }
 
 // CustomAssessmentAutomationsClientListByResourceGroupResponse contains the response from method CustomAssessmentAutomationsClient.ListByResourceGroup.
 type CustomAssessmentAutomationsClientListByResourceGroupResponse struct {
-	CustomAssessmentAutomationsClientListByResourceGroupResult
-	// RawResponse contains the underlying HTTP response.
-	RawResponse *http.Response
-}
-
-// CustomAssessmentAutomationsClientListByResourceGroupResult contains the result from method CustomAssessmentAutomationsClient.ListByResourceGroup.
-type CustomAssessmentAutomationsClientListByResourceGroupResult struct {
 	CustomAssessmentAutomationsListResult
 }
 
 // CustomAssessmentAutomationsClientListBySubscriptionResponse contains the response from method CustomAssessmentAutomationsClient.ListBySubscription.
 type CustomAssessmentAutomationsClientListBySubscriptionResponse struct {
-	CustomAssessmentAutomationsClientListBySubscriptionResult
-	// RawResponse contains the underlying HTTP response.
-	RawResponse *http.Response
-}
-
-// CustomAssessmentAutomationsClientListBySubscriptionResult contains the result from method CustomAssessmentAutomationsClient.ListBySubscription.
-type CustomAssessmentAutomationsClientListBySubscriptionResult struct {
 	CustomAssessmentAutomationsListResult
 }
 
 // CustomEntityStoreAssignmentsClientCreateResponse contains the response from method CustomEntityStoreAssignmentsClient.Create.
 type CustomEntityStoreAssignmentsClientCreateResponse struct {
-	CustomEntityStoreAssignmentsClientCreateResult
-	// RawResponse contains the underlying HTTP response.
-	RawResponse *http.Response
-}
-
-// CustomEntityStoreAssignmentsClientCreateResult contains the result from method CustomEntityStoreAssignmentsClient.Create.
-type CustomEntityStoreAssignmentsClientCreateResult struct {
 	CustomEntityStoreAssignment
 }
 
 // CustomEntityStoreAssignmentsClientDeleteResponse contains the response from method CustomEntityStoreAssignmentsClient.Delete.
 type CustomEntityStoreAssignmentsClientDeleteResponse struct {
-	// RawResponse contains the underlying HTTP response.
-	RawResponse *http.Response
+	// placeholder for future response values
 }
 
 // CustomEntityStoreAssignmentsClientGetResponse contains the response from method CustomEntityStoreAssignmentsClient.Get.
 type CustomEntityStoreAssignmentsClientGetResponse struct {
-	CustomEntityStoreAssignmentsClientGetResult
-	// RawResponse contains the underlying HTTP response.
-	RawResponse *http.Response
-}
-
-// CustomEntityStoreAssignmentsClientGetResult contains the result from method CustomEntityStoreAssignmentsClient.Get.
-type CustomEntityStoreAssignmentsClientGetResult struct {
 	CustomEntityStoreAssignment
 }
 
 // CustomEntityStoreAssignmentsClientListByResourceGroupResponse contains the response from method CustomEntityStoreAssignmentsClient.ListByResourceGroup.
 type CustomEntityStoreAssignmentsClientListByResourceGroupResponse struct {
-	CustomEntityStoreAssignmentsClientListByResourceGroupResult
-	// RawResponse contains the underlying HTTP response.
-	RawResponse *http.Response
-}
-
-// CustomEntityStoreAssignmentsClientListByResourceGroupResult contains the result from method CustomEntityStoreAssignmentsClient.ListByResourceGroup.
-type CustomEntityStoreAssignmentsClientListByResourceGroupResult struct {
 	CustomEntityStoreAssignmentsListResult
 }
 
 // CustomEntityStoreAssignmentsClientListBySubscriptionResponse contains the response from method CustomEntityStoreAssignmentsClient.ListBySubscription.
 type CustomEntityStoreAssignmentsClientListBySubscriptionResponse struct {
-	CustomEntityStoreAssignmentsClientListBySubscriptionResult
-	// RawResponse contains the underlying HTTP response.
-	RawResponse *http.Response
-}
-
-// CustomEntityStoreAssignmentsClientListBySubscriptionResult contains the result from method CustomEntityStoreAssignmentsClient.ListBySubscription.
-type CustomEntityStoreAssignmentsClientListBySubscriptionResult struct {
 	CustomEntityStoreAssignmentsListResult
 }
 
 // DeviceSecurityGroupsClientCreateOrUpdateResponse contains the response from method DeviceSecurityGroupsClient.CreateOrUpdate.
 type DeviceSecurityGroupsClientCreateOrUpdateResponse struct {
-	DeviceSecurityGroupsClientCreateOrUpdateResult
-	// RawResponse contains the underlying HTTP response.
-	RawResponse *http.Response
-}
-
-// DeviceSecurityGroupsClientCreateOrUpdateResult contains the result from method DeviceSecurityGroupsClient.CreateOrUpdate.
-type DeviceSecurityGroupsClientCreateOrUpdateResult struct {
 	DeviceSecurityGroup
 }
 
 // DeviceSecurityGroupsClientDeleteResponse contains the response from method DeviceSecurityGroupsClient.Delete.
 type DeviceSecurityGroupsClientDeleteResponse struct {
-	// RawResponse contains the underlying HTTP response.
-	RawResponse *http.Response
+	// placeholder for future response values
 }
 
 // DeviceSecurityGroupsClientGetResponse contains the response from method DeviceSecurityGroupsClient.Get.
 type DeviceSecurityGroupsClientGetResponse struct {
-	DeviceSecurityGroupsClientGetResult
-	// RawResponse contains the underlying HTTP response.
-	RawResponse *http.Response
-}
-
-// DeviceSecurityGroupsClientGetResult contains the result from method DeviceSecurityGroupsClient.Get.
-type DeviceSecurityGroupsClientGetResult struct {
 	DeviceSecurityGroup
 }
 
 // DeviceSecurityGroupsClientListResponse contains the response from method DeviceSecurityGroupsClient.List.
 type DeviceSecurityGroupsClientListResponse struct {
-	DeviceSecurityGroupsClientListResult
-	// RawResponse contains the underlying HTTP response.
-	RawResponse *http.Response
-}
-
-// DeviceSecurityGroupsClientListResult contains the result from method DeviceSecurityGroupsClient.List.
-type DeviceSecurityGroupsClientListResult struct {
 	DeviceSecurityGroupList
 }
 
 // DiscoveredSecuritySolutionsClientGetResponse contains the response from method DiscoveredSecuritySolutionsClient.Get.
 type DiscoveredSecuritySolutionsClientGetResponse struct {
-	DiscoveredSecuritySolutionsClientGetResult
-	// RawResponse contains the underlying HTTP response.
-	RawResponse *http.Response
-}
-
-// DiscoveredSecuritySolutionsClientGetResult contains the result from method DiscoveredSecuritySolutionsClient.Get.
-type DiscoveredSecuritySolutionsClientGetResult struct {
 	DiscoveredSecuritySolution
 }
 
 // DiscoveredSecuritySolutionsClientListByHomeRegionResponse contains the response from method DiscoveredSecuritySolutionsClient.ListByHomeRegion.
 type DiscoveredSecuritySolutionsClientListByHomeRegionResponse struct {
-	DiscoveredSecuritySolutionsClientListByHomeRegionResult
-	// RawResponse contains the underlying HTTP response.
-	RawResponse *http.Response
-}
-
-// DiscoveredSecuritySolutionsClientListByHomeRegionResult contains the result from method DiscoveredSecuritySolutionsClient.ListByHomeRegion.
-type DiscoveredSecuritySolutionsClientListByHomeRegionResult struct {
 	DiscoveredSecuritySolutionList
 }
 
 // DiscoveredSecuritySolutionsClientListResponse contains the response from method DiscoveredSecuritySolutionsClient.List.
 type DiscoveredSecuritySolutionsClientListResponse struct {
-	DiscoveredSecuritySolutionsClientListResult
-	// RawResponse contains the underlying HTTP response.
-	RawResponse *http.Response
-}
-
-// DiscoveredSecuritySolutionsClientListResult contains the result from method DiscoveredSecuritySolutionsClient.List.
-type DiscoveredSecuritySolutionsClientListResult struct {
 	DiscoveredSecuritySolutionList
 }
 
 // ExternalSecuritySolutionsClientGetResponse contains the response from method ExternalSecuritySolutionsClient.Get.
 type ExternalSecuritySolutionsClientGetResponse struct {
-	ExternalSecuritySolutionsClientGetResult
-	// RawResponse contains the underlying HTTP response.
-	RawResponse *http.Response
-}
-
-// ExternalSecuritySolutionsClientGetResult contains the result from method ExternalSecuritySolutionsClient.Get.
-type ExternalSecuritySolutionsClientGetResult struct {
 	ExternalSecuritySolution
 }
 
 // ExternalSecuritySolutionsClientListByHomeRegionResponse contains the response from method ExternalSecuritySolutionsClient.ListByHomeRegion.
 type ExternalSecuritySolutionsClientListByHomeRegionResponse struct {
-	ExternalSecuritySolutionsClientListByHomeRegionResult
-	// RawResponse contains the underlying HTTP response.
-	RawResponse *http.Response
-}
-
-// ExternalSecuritySolutionsClientListByHomeRegionResult contains the result from method ExternalSecuritySolutionsClient.ListByHomeRegion.
-type ExternalSecuritySolutionsClientListByHomeRegionResult struct {
 	ExternalSecuritySolutionList
 }
 
 // ExternalSecuritySolutionsClientListResponse contains the response from method ExternalSecuritySolutionsClient.List.
 type ExternalSecuritySolutionsClientListResponse struct {
-	ExternalSecuritySolutionsClientListResult
-	// RawResponse contains the underlying HTTP response.
-	RawResponse *http.Response
-}
-
-// ExternalSecuritySolutionsClientListResult contains the result from method ExternalSecuritySolutionsClient.List.
-type ExternalSecuritySolutionsClientListResult struct {
 	ExternalSecuritySolutionList
 }
 
 // InformationProtectionPoliciesClientCreateOrUpdateResponse contains the response from method InformationProtectionPoliciesClient.CreateOrUpdate.
 type InformationProtectionPoliciesClientCreateOrUpdateResponse struct {
-	InformationProtectionPoliciesClientCreateOrUpdateResult
-	// RawResponse contains the underlying HTTP response.
-	RawResponse *http.Response
-}
-
-// InformationProtectionPoliciesClientCreateOrUpdateResult contains the result from method InformationProtectionPoliciesClient.CreateOrUpdate.
-type InformationProtectionPoliciesClientCreateOrUpdateResult struct {
 	InformationProtectionPolicy
 }
 
 // InformationProtectionPoliciesClientGetResponse contains the response from method InformationProtectionPoliciesClient.Get.
 type InformationProtectionPoliciesClientGetResponse struct {
-	InformationProtectionPoliciesClientGetResult
-	// RawResponse contains the underlying HTTP response.
-	RawResponse *http.Response
-}
-
-// InformationProtectionPoliciesClientGetResult contains the result from method InformationProtectionPoliciesClient.Get.
-type InformationProtectionPoliciesClientGetResult struct {
 	InformationProtectionPolicy
 }
 
 // InformationProtectionPoliciesClientListResponse contains the response from method InformationProtectionPoliciesClient.List.
 type InformationProtectionPoliciesClientListResponse struct {
-	InformationProtectionPoliciesClientListResult
-	// RawResponse contains the underlying HTTP response.
-	RawResponse *http.Response
-}
-
-// InformationProtectionPoliciesClientListResult contains the result from method InformationProtectionPoliciesClient.List.
-type InformationProtectionPoliciesClientListResult struct {
 	InformationProtectionPolicyList
 }
 
 // IngestionSettingsClientCreateResponse contains the response from method IngestionSettingsClient.Create.
 type IngestionSettingsClientCreateResponse struct {
-	IngestionSettingsClientCreateResult
-	// RawResponse contains the underlying HTTP response.
-	RawResponse *http.Response
-}
-
-// IngestionSettingsClientCreateResult contains the result from method IngestionSettingsClient.Create.
-type IngestionSettingsClientCreateResult struct {
 	IngestionSetting
 }
 
 // IngestionSettingsClientDeleteResponse contains the response from method IngestionSettingsClient.Delete.
 type IngestionSettingsClientDeleteResponse struct {
-	// RawResponse contains the underlying HTTP response.
-	RawResponse *http.Response
+	// placeholder for future response values
 }
 
 // IngestionSettingsClientGetResponse contains the response from method IngestionSettingsClient.Get.
 type IngestionSettingsClientGetResponse struct {
-	IngestionSettingsClientGetResult
-	// RawResponse contains the underlying HTTP response.
-	RawResponse *http.Response
-}
-
-// IngestionSettingsClientGetResult contains the result from method IngestionSettingsClient.Get.
-type IngestionSettingsClientGetResult struct {
 	IngestionSetting
 }
 
 // IngestionSettingsClientListConnectionStringsResponse contains the response from method IngestionSettingsClient.ListConnectionStrings.
 type IngestionSettingsClientListConnectionStringsResponse struct {
-	IngestionSettingsClientListConnectionStringsResult
-	// RawResponse contains the underlying HTTP response.
-	RawResponse *http.Response
-}
-
-// IngestionSettingsClientListConnectionStringsResult contains the result from method IngestionSettingsClient.ListConnectionStrings.
-type IngestionSettingsClientListConnectionStringsResult struct {
 	ConnectionStrings
 }
 
 // IngestionSettingsClientListResponse contains the response from method IngestionSettingsClient.List.
 type IngestionSettingsClientListResponse struct {
-	IngestionSettingsClientListResult
-	// RawResponse contains the underlying HTTP response.
-	RawResponse *http.Response
-}
-
-// IngestionSettingsClientListResult contains the result from method IngestionSettingsClient.List.
-type IngestionSettingsClientListResult struct {
 	IngestionSettingList
 }
 
 // IngestionSettingsClientListTokensResponse contains the response from method IngestionSettingsClient.ListTokens.
 type IngestionSettingsClientListTokensResponse struct {
-	IngestionSettingsClientListTokensResult
-	// RawResponse contains the underlying HTTP response.
-	RawResponse *http.Response
-}
-
-// IngestionSettingsClientListTokensResult contains the result from method IngestionSettingsClient.ListTokens.
-type IngestionSettingsClientListTokensResult struct {
 	IngestionSettingToken
 }
 
 // IotSecuritySolutionAnalyticsClientGetResponse contains the response from method IotSecuritySolutionAnalyticsClient.Get.
 type IotSecuritySolutionAnalyticsClientGetResponse struct {
-	IotSecuritySolutionAnalyticsClientGetResult
-	// RawResponse contains the underlying HTTP response.
-	RawResponse *http.Response
-}
-
-// IotSecuritySolutionAnalyticsClientGetResult contains the result from method IotSecuritySolutionAnalyticsClient.Get.
-type IotSecuritySolutionAnalyticsClientGetResult struct {
 	IoTSecuritySolutionAnalyticsModel
 }
 
 // IotSecuritySolutionAnalyticsClientListResponse contains the response from method IotSecuritySolutionAnalyticsClient.List.
 type IotSecuritySolutionAnalyticsClientListResponse struct {
-	IotSecuritySolutionAnalyticsClientListResult
-	// RawResponse contains the underlying HTTP response.
-	RawResponse *http.Response
-}
-
-// IotSecuritySolutionAnalyticsClientListResult contains the result from method IotSecuritySolutionAnalyticsClient.List.
-type IotSecuritySolutionAnalyticsClientListResult struct {
 	IoTSecuritySolutionAnalyticsModelList
 }
 
 // IotSecuritySolutionClientCreateOrUpdateResponse contains the response from method IotSecuritySolutionClient.CreateOrUpdate.
 type IotSecuritySolutionClientCreateOrUpdateResponse struct {
-	IotSecuritySolutionClientCreateOrUpdateResult
-	// RawResponse contains the underlying HTTP response.
-	RawResponse *http.Response
-}
-
-// IotSecuritySolutionClientCreateOrUpdateResult contains the result from method IotSecuritySolutionClient.CreateOrUpdate.
-type IotSecuritySolutionClientCreateOrUpdateResult struct {
 	IoTSecuritySolutionModel
 }
 
 // IotSecuritySolutionClientDeleteResponse contains the response from method IotSecuritySolutionClient.Delete.
 type IotSecuritySolutionClientDeleteResponse struct {
-	// RawResponse contains the underlying HTTP response.
-	RawResponse *http.Response
+	// placeholder for future response values
 }
 
 // IotSecuritySolutionClientGetResponse contains the response from method IotSecuritySolutionClient.Get.
 type IotSecuritySolutionClientGetResponse struct {
-	IotSecuritySolutionClientGetResult
-	// RawResponse contains the underlying HTTP response.
-	RawResponse *http.Response
-}
-
-// IotSecuritySolutionClientGetResult contains the result from method IotSecuritySolutionClient.Get.
-type IotSecuritySolutionClientGetResult struct {
 	IoTSecuritySolutionModel
 }
 
 // IotSecuritySolutionClientListByResourceGroupResponse contains the response from method IotSecuritySolutionClient.ListByResourceGroup.
 type IotSecuritySolutionClientListByResourceGroupResponse struct {
-	IotSecuritySolutionClientListByResourceGroupResult
-	// RawResponse contains the underlying HTTP response.
-	RawResponse *http.Response
-}
-
-// IotSecuritySolutionClientListByResourceGroupResult contains the result from method IotSecuritySolutionClient.ListByResourceGroup.
-type IotSecuritySolutionClientListByResourceGroupResult struct {
 	IoTSecuritySolutionsList
 }
 
 // IotSecuritySolutionClientListBySubscriptionResponse contains the response from method IotSecuritySolutionClient.ListBySubscription.
 type IotSecuritySolutionClientListBySubscriptionResponse struct {
-	IotSecuritySolutionClientListBySubscriptionResult
-	// RawResponse contains the underlying HTTP response.
-	RawResponse *http.Response
-}
-
-// IotSecuritySolutionClientListBySubscriptionResult contains the result from method IotSecuritySolutionClient.ListBySubscription.
-type IotSecuritySolutionClientListBySubscriptionResult struct {
 	IoTSecuritySolutionsList
 }
 
 // IotSecuritySolutionClientUpdateResponse contains the response from method IotSecuritySolutionClient.Update.
 type IotSecuritySolutionClientUpdateResponse struct {
-	IotSecuritySolutionClientUpdateResult
-	// RawResponse contains the underlying HTTP response.
-	RawResponse *http.Response
-}
-
-// IotSecuritySolutionClientUpdateResult contains the result from method IotSecuritySolutionClient.Update.
-type IotSecuritySolutionClientUpdateResult struct {
 	IoTSecuritySolutionModel
 }
 
 // IotSecuritySolutionsAnalyticsAggregatedAlertClientDismissResponse contains the response from method IotSecuritySolutionsAnalyticsAggregatedAlertClient.Dismiss.
 type IotSecuritySolutionsAnalyticsAggregatedAlertClientDismissResponse struct {
-	// RawResponse contains the underlying HTTP response.
-	RawResponse *http.Response
+	// placeholder for future response values
 }
 
 // IotSecuritySolutionsAnalyticsAggregatedAlertClientGetResponse contains the response from method IotSecuritySolutionsAnalyticsAggregatedAlertClient.Get.
 type IotSecuritySolutionsAnalyticsAggregatedAlertClientGetResponse struct {
-	IotSecuritySolutionsAnalyticsAggregatedAlertClientGetResult
-	// RawResponse contains the underlying HTTP response.
-	RawResponse *http.Response
-}
-
-// IotSecuritySolutionsAnalyticsAggregatedAlertClientGetResult contains the result from method IotSecuritySolutionsAnalyticsAggregatedAlertClient.Get.
-type IotSecuritySolutionsAnalyticsAggregatedAlertClientGetResult struct {
 	IoTSecurityAggregatedAlert
 }
 
 // IotSecuritySolutionsAnalyticsAggregatedAlertClientListResponse contains the response from method IotSecuritySolutionsAnalyticsAggregatedAlertClient.List.
 type IotSecuritySolutionsAnalyticsAggregatedAlertClientListResponse struct {
-	IotSecuritySolutionsAnalyticsAggregatedAlertClientListResult
-	// RawResponse contains the underlying HTTP response.
-	RawResponse *http.Response
-}
-
-// IotSecuritySolutionsAnalyticsAggregatedAlertClientListResult contains the result from method IotSecuritySolutionsAnalyticsAggregatedAlertClient.List.
-type IotSecuritySolutionsAnalyticsAggregatedAlertClientListResult struct {
 	IoTSecurityAggregatedAlertList
 }
 
 // IotSecuritySolutionsAnalyticsRecommendationClientGetResponse contains the response from method IotSecuritySolutionsAnalyticsRecommendationClient.Get.
 type IotSecuritySolutionsAnalyticsRecommendationClientGetResponse struct {
-	IotSecuritySolutionsAnalyticsRecommendationClientGetResult
-	// RawResponse contains the underlying HTTP response.
-	RawResponse *http.Response
-}
-
-// IotSecuritySolutionsAnalyticsRecommendationClientGetResult contains the result from method IotSecuritySolutionsAnalyticsRecommendationClient.Get.
-type IotSecuritySolutionsAnalyticsRecommendationClientGetResult struct {
 	IoTSecurityAggregatedRecommendation
 }
 
 // IotSecuritySolutionsAnalyticsRecommendationClientListResponse contains the response from method IotSecuritySolutionsAnalyticsRecommendationClient.List.
 type IotSecuritySolutionsAnalyticsRecommendationClientListResponse struct {
-	IotSecuritySolutionsAnalyticsRecommendationClientListResult
-	// RawResponse contains the underlying HTTP response.
-	RawResponse *http.Response
-}
-
-// IotSecuritySolutionsAnalyticsRecommendationClientListResult contains the result from method IotSecuritySolutionsAnalyticsRecommendationClient.List.
-type IotSecuritySolutionsAnalyticsRecommendationClientListResult struct {
 	IoTSecurityAggregatedRecommendationList
 }
 
 // JitNetworkAccessPoliciesClientCreateOrUpdateResponse contains the response from method JitNetworkAccessPoliciesClient.CreateOrUpdate.
 type JitNetworkAccessPoliciesClientCreateOrUpdateResponse struct {
-	JitNetworkAccessPoliciesClientCreateOrUpdateResult
-	// RawResponse contains the underlying HTTP response.
-	RawResponse *http.Response
-}
-
-// JitNetworkAccessPoliciesClientCreateOrUpdateResult contains the result from method JitNetworkAccessPoliciesClient.CreateOrUpdate.
-type JitNetworkAccessPoliciesClientCreateOrUpdateResult struct {
 	JitNetworkAccessPolicy
 }
 
 // JitNetworkAccessPoliciesClientDeleteResponse contains the response from method JitNetworkAccessPoliciesClient.Delete.
 type JitNetworkAccessPoliciesClientDeleteResponse struct {
-	// RawResponse contains the underlying HTTP response.
-	RawResponse *http.Response
+	// placeholder for future response values
 }
 
 // JitNetworkAccessPoliciesClientGetResponse contains the response from method JitNetworkAccessPoliciesClient.Get.
 type JitNetworkAccessPoliciesClientGetResponse struct {
-	JitNetworkAccessPoliciesClientGetResult
-	// RawResponse contains the underlying HTTP response.
-	RawResponse *http.Response
-}
-
-// JitNetworkAccessPoliciesClientGetResult contains the result from method JitNetworkAccessPoliciesClient.Get.
-type JitNetworkAccessPoliciesClientGetResult struct {
 	JitNetworkAccessPolicy
 }
 
 // JitNetworkAccessPoliciesClientInitiateResponse contains the response from method JitNetworkAccessPoliciesClient.Initiate.
 type JitNetworkAccessPoliciesClientInitiateResponse struct {
-	JitNetworkAccessPoliciesClientInitiateResult
-	// RawResponse contains the underlying HTTP response.
-	RawResponse *http.Response
-}
-
-// JitNetworkAccessPoliciesClientInitiateResult contains the result from method JitNetworkAccessPoliciesClient.Initiate.
-type JitNetworkAccessPoliciesClientInitiateResult struct {
 	JitNetworkAccessRequest
 }
 
 // JitNetworkAccessPoliciesClientListByRegionResponse contains the response from method JitNetworkAccessPoliciesClient.ListByRegion.
 type JitNetworkAccessPoliciesClientListByRegionResponse struct {
-	JitNetworkAccessPoliciesClientListByRegionResult
-	// RawResponse contains the underlying HTTP response.
-	RawResponse *http.Response
-}
-
-// JitNetworkAccessPoliciesClientListByRegionResult contains the result from method JitNetworkAccessPoliciesClient.ListByRegion.
-type JitNetworkAccessPoliciesClientListByRegionResult struct {
 	JitNetworkAccessPoliciesList
 }
 
 // JitNetworkAccessPoliciesClientListByResourceGroupAndRegionResponse contains the response from method JitNetworkAccessPoliciesClient.ListByResourceGroupAndRegion.
 type JitNetworkAccessPoliciesClientListByResourceGroupAndRegionResponse struct {
-	JitNetworkAccessPoliciesClientListByResourceGroupAndRegionResult
-	// RawResponse contains the underlying HTTP response.
-	RawResponse *http.Response
-}
-
-// JitNetworkAccessPoliciesClientListByResourceGroupAndRegionResult contains the result from method JitNetworkAccessPoliciesClient.ListByResourceGroupAndRegion.
-type JitNetworkAccessPoliciesClientListByResourceGroupAndRegionResult struct {
 	JitNetworkAccessPoliciesList
 }
 
 // JitNetworkAccessPoliciesClientListByResourceGroupResponse contains the response from method JitNetworkAccessPoliciesClient.ListByResourceGroup.
 type JitNetworkAccessPoliciesClientListByResourceGroupResponse struct {
-	JitNetworkAccessPoliciesClientListByResourceGroupResult
-	// RawResponse contains the underlying HTTP response.
-	RawResponse *http.Response
-}
-
-// JitNetworkAccessPoliciesClientListByResourceGroupResult contains the result from method JitNetworkAccessPoliciesClient.ListByResourceGroup.
-type JitNetworkAccessPoliciesClientListByResourceGroupResult struct {
 	JitNetworkAccessPoliciesList
 }
 
 // JitNetworkAccessPoliciesClientListResponse contains the response from method JitNetworkAccessPoliciesClient.List.
 type JitNetworkAccessPoliciesClientListResponse struct {
-	JitNetworkAccessPoliciesClientListResult
-	// RawResponse contains the underlying HTTP response.
-	RawResponse *http.Response
-}
-
-// JitNetworkAccessPoliciesClientListResult contains the result from method JitNetworkAccessPoliciesClient.List.
-type JitNetworkAccessPoliciesClientListResult struct {
 	JitNetworkAccessPoliciesList
 }
 
 // LocationsClientGetResponse contains the response from method LocationsClient.Get.
 type LocationsClientGetResponse struct {
-	LocationsClientGetResult
-	// RawResponse contains the underlying HTTP response.
-	RawResponse *http.Response
-}
-
-// LocationsClientGetResult contains the result from method LocationsClient.Get.
-type LocationsClientGetResult struct {
 	AscLocation
 }
 
 // LocationsClientListResponse contains the response from method LocationsClient.List.
 type LocationsClientListResponse struct {
-	LocationsClientListResult
-	// RawResponse contains the underlying HTTP response.
-	RawResponse *http.Response
-}
-
-// LocationsClientListResult contains the result from method LocationsClient.List.
-type LocationsClientListResult struct {
 	AscLocationList
 }
 
 // MdeOnboardingsClientGetResponse contains the response from method MdeOnboardingsClient.Get.
 type MdeOnboardingsClientGetResponse struct {
-	MdeOnboardingsClientGetResult
-	// RawResponse contains the underlying HTTP response.
-	RawResponse *http.Response
-}
-
-// MdeOnboardingsClientGetResult contains the result from method MdeOnboardingsClient.Get.
-type MdeOnboardingsClientGetResult struct {
 	MdeOnboardingData
 }
 
 // MdeOnboardingsClientListResponse contains the response from method MdeOnboardingsClient.List.
 type MdeOnboardingsClientListResponse struct {
-	MdeOnboardingsClientListResult
-	// RawResponse contains the underlying HTTP response.
-	RawResponse *http.Response
-}
-
-// MdeOnboardingsClientListResult contains the result from method MdeOnboardingsClient.List.
-type MdeOnboardingsClientListResult struct {
 	MdeOnboardingDataList
 }
 
 // OperationsClientListResponse contains the response from method OperationsClient.List.
 type OperationsClientListResponse struct {
-	OperationsClientListResult
-	// RawResponse contains the underlying HTTP response.
-	RawResponse *http.Response
-}
-
-// OperationsClientListResult contains the result from method OperationsClient.List.
-type OperationsClientListResult struct {
 	OperationList
 }
 
 // PricingsClientGetResponse contains the response from method PricingsClient.Get.
 type PricingsClientGetResponse struct {
-	PricingsClientGetResult
-	// RawResponse contains the underlying HTTP response.
-	RawResponse *http.Response
-}
-
-// PricingsClientGetResult contains the result from method PricingsClient.Get.
-type PricingsClientGetResult struct {
 	Pricing
 }
 
 // PricingsClientListResponse contains the response from method PricingsClient.List.
 type PricingsClientListResponse struct {
-	PricingsClientListResult
-	// RawResponse contains the underlying HTTP response.
-	RawResponse *http.Response
-}
-
-// PricingsClientListResult contains the result from method PricingsClient.List.
-type PricingsClientListResult struct {
 	PricingList
 }
 
 // PricingsClientUpdateResponse contains the response from method PricingsClient.Update.
 type PricingsClientUpdateResponse struct {
-	PricingsClientUpdateResult
-	// RawResponse contains the underlying HTTP response.
-	RawResponse *http.Response
-}
-
-// PricingsClientUpdateResult contains the result from method PricingsClient.Update.
-type PricingsClientUpdateResult struct {
 	Pricing
 }
 
 // RegulatoryComplianceAssessmentsClientGetResponse contains the response from method RegulatoryComplianceAssessmentsClient.Get.
 type RegulatoryComplianceAssessmentsClientGetResponse struct {
-	RegulatoryComplianceAssessmentsClientGetResult
-	// RawResponse contains the underlying HTTP response.
-	RawResponse *http.Response
-}
-
-// RegulatoryComplianceAssessmentsClientGetResult contains the result from method RegulatoryComplianceAssessmentsClient.Get.
-type RegulatoryComplianceAssessmentsClientGetResult struct {
 	RegulatoryComplianceAssessment
 }
 
 // RegulatoryComplianceAssessmentsClientListResponse contains the response from method RegulatoryComplianceAssessmentsClient.List.
 type RegulatoryComplianceAssessmentsClientListResponse struct {
-	RegulatoryComplianceAssessmentsClientListResult
-	// RawResponse contains the underlying HTTP response.
-	RawResponse *http.Response
-}
-
-// RegulatoryComplianceAssessmentsClientListResult contains the result from method RegulatoryComplianceAssessmentsClient.List.
-type RegulatoryComplianceAssessmentsClientListResult struct {
 	RegulatoryComplianceAssessmentList
 }
 
 // RegulatoryComplianceControlsClientGetResponse contains the response from method RegulatoryComplianceControlsClient.Get.
 type RegulatoryComplianceControlsClientGetResponse struct {
-	RegulatoryComplianceControlsClientGetResult
-	// RawResponse contains the underlying HTTP response.
-	RawResponse *http.Response
-}
-
-// RegulatoryComplianceControlsClientGetResult contains the result from method RegulatoryComplianceControlsClient.Get.
-type RegulatoryComplianceControlsClientGetResult struct {
 	RegulatoryComplianceControl
 }
 
 // RegulatoryComplianceControlsClientListResponse contains the response from method RegulatoryComplianceControlsClient.List.
 type RegulatoryComplianceControlsClientListResponse struct {
-	RegulatoryComplianceControlsClientListResult
-	// RawResponse contains the underlying HTTP response.
-	RawResponse *http.Response
-}
-
-// RegulatoryComplianceControlsClientListResult contains the result from method RegulatoryComplianceControlsClient.List.
-type RegulatoryComplianceControlsClientListResult struct {
 	RegulatoryComplianceControlList
 }
 
 // RegulatoryComplianceStandardsClientGetResponse contains the response from method RegulatoryComplianceStandardsClient.Get.
 type RegulatoryComplianceStandardsClientGetResponse struct {
-	RegulatoryComplianceStandardsClientGetResult
-	// RawResponse contains the underlying HTTP response.
-	RawResponse *http.Response
-}
-
-// RegulatoryComplianceStandardsClientGetResult contains the result from method RegulatoryComplianceStandardsClient.Get.
-type RegulatoryComplianceStandardsClientGetResult struct {
 	RegulatoryComplianceStandard
 }
 
 // RegulatoryComplianceStandardsClientListResponse contains the response from method RegulatoryComplianceStandardsClient.List.
 type RegulatoryComplianceStandardsClientListResponse struct {
-	RegulatoryComplianceStandardsClientListResult
-	// RawResponse contains the underlying HTTP response.
-	RawResponse *http.Response
-}
-
-// RegulatoryComplianceStandardsClientListResult contains the result from method RegulatoryComplianceStandardsClient.List.
-type RegulatoryComplianceStandardsClientListResult struct {
 	RegulatoryComplianceStandardList
 }
 
 // SQLVulnerabilityAssessmentBaselineRulesClientAddResponse contains the response from method SQLVulnerabilityAssessmentBaselineRulesClient.Add.
 type SQLVulnerabilityAssessmentBaselineRulesClientAddResponse struct {
-	SQLVulnerabilityAssessmentBaselineRulesClientAddResult
-	// RawResponse contains the underlying HTTP response.
-	RawResponse *http.Response
-}
-
-// SQLVulnerabilityAssessmentBaselineRulesClientAddResult contains the result from method SQLVulnerabilityAssessmentBaselineRulesClient.Add.
-type SQLVulnerabilityAssessmentBaselineRulesClientAddResult struct {
 	RulesResults
 }
 
 // SQLVulnerabilityAssessmentBaselineRulesClientCreateOrUpdateResponse contains the response from method SQLVulnerabilityAssessmentBaselineRulesClient.CreateOrUpdate.
 type SQLVulnerabilityAssessmentBaselineRulesClientCreateOrUpdateResponse struct {
-	SQLVulnerabilityAssessmentBaselineRulesClientCreateOrUpdateResult
-	// RawResponse contains the underlying HTTP response.
-	RawResponse *http.Response
-}
-
-// SQLVulnerabilityAssessmentBaselineRulesClientCreateOrUpdateResult contains the result from method SQLVulnerabilityAssessmentBaselineRulesClient.CreateOrUpdate.
-type SQLVulnerabilityAssessmentBaselineRulesClientCreateOrUpdateResult struct {
 	RuleResults
 }
 
 // SQLVulnerabilityAssessmentBaselineRulesClientDeleteResponse contains the response from method SQLVulnerabilityAssessmentBaselineRulesClient.Delete.
 type SQLVulnerabilityAssessmentBaselineRulesClientDeleteResponse struct {
-	// RawResponse contains the underlying HTTP response.
-	RawResponse *http.Response
+	// placeholder for future response values
 }
 
 // SQLVulnerabilityAssessmentBaselineRulesClientGetResponse contains the response from method SQLVulnerabilityAssessmentBaselineRulesClient.Get.
 type SQLVulnerabilityAssessmentBaselineRulesClientGetResponse struct {
-	SQLVulnerabilityAssessmentBaselineRulesClientGetResult
-	// RawResponse contains the underlying HTTP response.
-	RawResponse *http.Response
-}
-
-// SQLVulnerabilityAssessmentBaselineRulesClientGetResult contains the result from method SQLVulnerabilityAssessmentBaselineRulesClient.Get.
-type SQLVulnerabilityAssessmentBaselineRulesClientGetResult struct {
 	RuleResults
 }
 
 // SQLVulnerabilityAssessmentBaselineRulesClientListResponse contains the response from method SQLVulnerabilityAssessmentBaselineRulesClient.List.
 type SQLVulnerabilityAssessmentBaselineRulesClientListResponse struct {
-	SQLVulnerabilityAssessmentBaselineRulesClientListResult
-	// RawResponse contains the underlying HTTP response.
-	RawResponse *http.Response
-}
-
-// SQLVulnerabilityAssessmentBaselineRulesClientListResult contains the result from method SQLVulnerabilityAssessmentBaselineRulesClient.List.
-type SQLVulnerabilityAssessmentBaselineRulesClientListResult struct {
 	RulesResults
 }
 
 // SQLVulnerabilityAssessmentScanResultsClientGetResponse contains the response from method SQLVulnerabilityAssessmentScanResultsClient.Get.
 type SQLVulnerabilityAssessmentScanResultsClientGetResponse struct {
-	SQLVulnerabilityAssessmentScanResultsClientGetResult
-	// RawResponse contains the underlying HTTP response.
-	RawResponse *http.Response
-}
-
-// SQLVulnerabilityAssessmentScanResultsClientGetResult contains the result from method SQLVulnerabilityAssessmentScanResultsClient.Get.
-type SQLVulnerabilityAssessmentScanResultsClientGetResult struct {
 	ScanResult
 }
 
 // SQLVulnerabilityAssessmentScanResultsClientListResponse contains the response from method SQLVulnerabilityAssessmentScanResultsClient.List.
 type SQLVulnerabilityAssessmentScanResultsClientListResponse struct {
-	SQLVulnerabilityAssessmentScanResultsClientListResult
-	// RawResponse contains the underlying HTTP response.
-	RawResponse *http.Response
-}
-
-// SQLVulnerabilityAssessmentScanResultsClientListResult contains the result from method SQLVulnerabilityAssessmentScanResultsClient.List.
-type SQLVulnerabilityAssessmentScanResultsClientListResult struct {
 	ScanResults
 }
 
 // SQLVulnerabilityAssessmentScansClientGetResponse contains the response from method SQLVulnerabilityAssessmentScansClient.Get.
 type SQLVulnerabilityAssessmentScansClientGetResponse struct {
-	SQLVulnerabilityAssessmentScansClientGetResult
-	// RawResponse contains the underlying HTTP response.
-	RawResponse *http.Response
-}
-
-// SQLVulnerabilityAssessmentScansClientGetResult contains the result from method SQLVulnerabilityAssessmentScansClient.Get.
-type SQLVulnerabilityAssessmentScansClientGetResult struct {
 	Scan
 }
 
 // SQLVulnerabilityAssessmentScansClientListResponse contains the response from method SQLVulnerabilityAssessmentScansClient.List.
 type SQLVulnerabilityAssessmentScansClientListResponse struct {
-	SQLVulnerabilityAssessmentScansClientListResult
-	// RawResponse contains the underlying HTTP response.
-	RawResponse *http.Response
-}
-
-// SQLVulnerabilityAssessmentScansClientListResult contains the result from method SQLVulnerabilityAssessmentScansClient.List.
-type SQLVulnerabilityAssessmentScansClientListResult struct {
 	Scans
 }
 
 // SecureScoreControlDefinitionsClientListBySubscriptionResponse contains the response from method SecureScoreControlDefinitionsClient.ListBySubscription.
 type SecureScoreControlDefinitionsClientListBySubscriptionResponse struct {
-	SecureScoreControlDefinitionsClientListBySubscriptionResult
-	// RawResponse contains the underlying HTTP response.
-	RawResponse *http.Response
-}
-
-// SecureScoreControlDefinitionsClientListBySubscriptionResult contains the result from method SecureScoreControlDefinitionsClient.ListBySubscription.
-type SecureScoreControlDefinitionsClientListBySubscriptionResult struct {
 	SecureScoreControlDefinitionList
 }
 
 // SecureScoreControlDefinitionsClientListResponse contains the response from method SecureScoreControlDefinitionsClient.List.
 type SecureScoreControlDefinitionsClientListResponse struct {
-	SecureScoreControlDefinitionsClientListResult
-	// RawResponse contains the underlying HTTP response.
-	RawResponse *http.Response
-}
-
-// SecureScoreControlDefinitionsClientListResult contains the result from method SecureScoreControlDefinitionsClient.List.
-type SecureScoreControlDefinitionsClientListResult struct {
 	SecureScoreControlDefinitionList
 }
 
 // SecureScoreControlsClientListBySecureScoreResponse contains the response from method SecureScoreControlsClient.ListBySecureScore.
 type SecureScoreControlsClientListBySecureScoreResponse struct {
-	SecureScoreControlsClientListBySecureScoreResult
-	// RawResponse contains the underlying HTTP response.
-	RawResponse *http.Response
-}
-
-// SecureScoreControlsClientListBySecureScoreResult contains the result from method SecureScoreControlsClient.ListBySecureScore.
-type SecureScoreControlsClientListBySecureScoreResult struct {
 	SecureScoreControlList
 }
 
 // SecureScoreControlsClientListResponse contains the response from method SecureScoreControlsClient.List.
 type SecureScoreControlsClientListResponse struct {
-	SecureScoreControlsClientListResult
-	// RawResponse contains the underlying HTTP response.
-	RawResponse *http.Response
-}
-
-// SecureScoreControlsClientListResult contains the result from method SecureScoreControlsClient.List.
-type SecureScoreControlsClientListResult struct {
 	SecureScoreControlList
 }
 
 // SecureScoresClientGetResponse contains the response from method SecureScoresClient.Get.
 type SecureScoresClientGetResponse struct {
-	SecureScoresClientGetResult
-	// RawResponse contains the underlying HTTP response.
-	RawResponse *http.Response
-}
-
-// SecureScoresClientGetResult contains the result from method SecureScoresClient.Get.
-type SecureScoresClientGetResult struct {
 	SecureScoreItem
 }
 
 // SecureScoresClientListResponse contains the response from method SecureScoresClient.List.
 type SecureScoresClientListResponse struct {
-	SecureScoresClientListResult
-	// RawResponse contains the underlying HTTP response.
-	RawResponse *http.Response
-}
-
-// SecureScoresClientListResult contains the result from method SecureScoresClient.List.
-type SecureScoresClientListResult struct {
 	SecureScoresList
 }
 
 // ServerVulnerabilityAssessmentClientCreateOrUpdateResponse contains the response from method ServerVulnerabilityAssessmentClient.CreateOrUpdate.
 type ServerVulnerabilityAssessmentClientCreateOrUpdateResponse struct {
-	ServerVulnerabilityAssessmentClientCreateOrUpdateResult
-	// RawResponse contains the underlying HTTP response.
-	RawResponse *http.Response
-}
-
-// ServerVulnerabilityAssessmentClientCreateOrUpdateResult contains the result from method ServerVulnerabilityAssessmentClient.CreateOrUpdate.
-type ServerVulnerabilityAssessmentClientCreateOrUpdateResult struct {
 	ServerVulnerabilityAssessment
-}
-
-// ServerVulnerabilityAssessmentClientDeletePollerResponse contains the response from method ServerVulnerabilityAssessmentClient.Delete.
-type ServerVulnerabilityAssessmentClientDeletePollerResponse struct {
-	// Poller contains an initialized poller.
-	Poller *ServerVulnerabilityAssessmentClientDeletePoller
-
-	// RawResponse contains the underlying HTTP response.
-	RawResponse *http.Response
-}
-
-// PollUntilDone will poll the service endpoint until a terminal state is reached or an error is received.
-// freq: the time to wait between intervals in absence of a Retry-After header. Allowed minimum is one second.
-// A good starting value is 30 seconds. Note that some resources might benefit from a different value.
-func (l ServerVulnerabilityAssessmentClientDeletePollerResponse) PollUntilDone(ctx context.Context, freq time.Duration) (ServerVulnerabilityAssessmentClientDeleteResponse, error) {
-	respType := ServerVulnerabilityAssessmentClientDeleteResponse{}
-	resp, err := l.Poller.pt.PollUntilDone(ctx, freq, nil)
-	if err != nil {
-		return respType, err
-	}
-	respType.RawResponse = resp
-	return respType, nil
-}
-
-// Resume rehydrates a ServerVulnerabilityAssessmentClientDeletePollerResponse from the provided client and resume token.
-func (l *ServerVulnerabilityAssessmentClientDeletePollerResponse) Resume(ctx context.Context, client *ServerVulnerabilityAssessmentClient, token string) error {
-	pt, err := armruntime.NewPollerFromResumeToken("ServerVulnerabilityAssessmentClient.Delete", token, client.pl)
-	if err != nil {
-		return err
-	}
-	poller := &ServerVulnerabilityAssessmentClientDeletePoller{
-		pt: pt,
-	}
-	resp, err := poller.Poll(ctx)
-	if err != nil {
-		return err
-	}
-	l.Poller = poller
-	l.RawResponse = resp
-	return nil
 }
 
 // ServerVulnerabilityAssessmentClientDeleteResponse contains the response from method ServerVulnerabilityAssessmentClient.Delete.
 type ServerVulnerabilityAssessmentClientDeleteResponse struct {
-	// RawResponse contains the underlying HTTP response.
-	RawResponse *http.Response
+	// placeholder for future response values
 }
 
 // ServerVulnerabilityAssessmentClientGetResponse contains the response from method ServerVulnerabilityAssessmentClient.Get.
 type ServerVulnerabilityAssessmentClientGetResponse struct {
-	ServerVulnerabilityAssessmentClientGetResult
-	// RawResponse contains the underlying HTTP response.
-	RawResponse *http.Response
-}
-
-// ServerVulnerabilityAssessmentClientGetResult contains the result from method ServerVulnerabilityAssessmentClient.Get.
-type ServerVulnerabilityAssessmentClientGetResult struct {
 	ServerVulnerabilityAssessment
 }
 
 // ServerVulnerabilityAssessmentClientListByExtendedResourceResponse contains the response from method ServerVulnerabilityAssessmentClient.ListByExtendedResource.
 type ServerVulnerabilityAssessmentClientListByExtendedResourceResponse struct {
-	ServerVulnerabilityAssessmentClientListByExtendedResourceResult
-	// RawResponse contains the underlying HTTP response.
-	RawResponse *http.Response
-}
-
-// ServerVulnerabilityAssessmentClientListByExtendedResourceResult contains the result from method ServerVulnerabilityAssessmentClient.ListByExtendedResource.
-type ServerVulnerabilityAssessmentClientListByExtendedResourceResult struct {
 	ServerVulnerabilityAssessmentsList
 }
 
 // SettingsClientGetResponse contains the response from method SettingsClient.Get.
 type SettingsClientGetResponse struct {
-	SettingsClientGetResult
-	// RawResponse contains the underlying HTTP response.
-	RawResponse *http.Response
-}
-
-// SettingsClientGetResult contains the result from method SettingsClient.Get.
-type SettingsClientGetResult struct {
 	SettingClassification
 }
 
-// UnmarshalJSON implements the json.Unmarshaller interface for type SettingsClientGetResult.
-func (s *SettingsClientGetResult) UnmarshalJSON(data []byte) error {
+// UnmarshalJSON implements the json.Unmarshaller interface for type SettingsClientGetResponse.
+func (s *SettingsClientGetResponse) UnmarshalJSON(data []byte) error {
 	res, err := unmarshalSettingClassification(data)
 	if err != nil {
 		return err
@@ -1809,30 +775,16 @@ func (s *SettingsClientGetResult) UnmarshalJSON(data []byte) error {
 
 // SettingsClientListResponse contains the response from method SettingsClient.List.
 type SettingsClientListResponse struct {
-	SettingsClientListResult
-	// RawResponse contains the underlying HTTP response.
-	RawResponse *http.Response
-}
-
-// SettingsClientListResult contains the result from method SettingsClient.List.
-type SettingsClientListResult struct {
 	SettingsList
 }
 
 // SettingsClientUpdateResponse contains the response from method SettingsClient.Update.
 type SettingsClientUpdateResponse struct {
-	SettingsClientUpdateResult
-	// RawResponse contains the underlying HTTP response.
-	RawResponse *http.Response
-}
-
-// SettingsClientUpdateResult contains the result from method SettingsClient.Update.
-type SettingsClientUpdateResult struct {
 	SettingClassification
 }
 
-// UnmarshalJSON implements the json.Unmarshaller interface for type SettingsClientUpdateResult.
-func (s *SettingsClientUpdateResult) UnmarshalJSON(data []byte) error {
+// UnmarshalJSON implements the json.Unmarshaller interface for type SettingsClientUpdateResponse.
+func (s *SettingsClientUpdateResponse) UnmarshalJSON(data []byte) error {
 	res, err := unmarshalSettingClassification(data)
 	if err != nil {
 		return err
@@ -1843,282 +795,125 @@ func (s *SettingsClientUpdateResult) UnmarshalJSON(data []byte) error {
 
 // SoftwareInventoriesClientGetResponse contains the response from method SoftwareInventoriesClient.Get.
 type SoftwareInventoriesClientGetResponse struct {
-	SoftwareInventoriesClientGetResult
-	// RawResponse contains the underlying HTTP response.
-	RawResponse *http.Response
-}
-
-// SoftwareInventoriesClientGetResult contains the result from method SoftwareInventoriesClient.Get.
-type SoftwareInventoriesClientGetResult struct {
 	Software
 }
 
 // SoftwareInventoriesClientListByExtendedResourceResponse contains the response from method SoftwareInventoriesClient.ListByExtendedResource.
 type SoftwareInventoriesClientListByExtendedResourceResponse struct {
-	SoftwareInventoriesClientListByExtendedResourceResult
-	// RawResponse contains the underlying HTTP response.
-	RawResponse *http.Response
-}
-
-// SoftwareInventoriesClientListByExtendedResourceResult contains the result from method SoftwareInventoriesClient.ListByExtendedResource.
-type SoftwareInventoriesClientListByExtendedResourceResult struct {
 	SoftwaresList
 }
 
 // SoftwareInventoriesClientListBySubscriptionResponse contains the response from method SoftwareInventoriesClient.ListBySubscription.
 type SoftwareInventoriesClientListBySubscriptionResponse struct {
-	SoftwareInventoriesClientListBySubscriptionResult
-	// RawResponse contains the underlying HTTP response.
-	RawResponse *http.Response
-}
-
-// SoftwareInventoriesClientListBySubscriptionResult contains the result from method SoftwareInventoriesClient.ListBySubscription.
-type SoftwareInventoriesClientListBySubscriptionResult struct {
 	SoftwaresList
 }
 
 // SolutionsClientGetResponse contains the response from method SolutionsClient.Get.
 type SolutionsClientGetResponse struct {
-	SolutionsClientGetResult
-	// RawResponse contains the underlying HTTP response.
-	RawResponse *http.Response
-}
-
-// SolutionsClientGetResult contains the result from method SolutionsClient.Get.
-type SolutionsClientGetResult struct {
 	Solution
 }
 
 // SolutionsClientListResponse contains the response from method SolutionsClient.List.
 type SolutionsClientListResponse struct {
-	SolutionsClientListResult
-	// RawResponse contains the underlying HTTP response.
-	RawResponse *http.Response
-}
-
-// SolutionsClientListResult contains the result from method SolutionsClient.List.
-type SolutionsClientListResult struct {
 	SolutionList
 }
 
 // SolutionsReferenceDataClientListByHomeRegionResponse contains the response from method SolutionsReferenceDataClient.ListByHomeRegion.
 type SolutionsReferenceDataClientListByHomeRegionResponse struct {
-	SolutionsReferenceDataClientListByHomeRegionResult
-	// RawResponse contains the underlying HTTP response.
-	RawResponse *http.Response
-}
-
-// SolutionsReferenceDataClientListByHomeRegionResult contains the result from method SolutionsReferenceDataClient.ListByHomeRegion.
-type SolutionsReferenceDataClientListByHomeRegionResult struct {
 	SolutionsReferenceDataList
 }
 
 // SolutionsReferenceDataClientListResponse contains the response from method SolutionsReferenceDataClient.List.
 type SolutionsReferenceDataClientListResponse struct {
-	SolutionsReferenceDataClientListResult
-	// RawResponse contains the underlying HTTP response.
-	RawResponse *http.Response
-}
-
-// SolutionsReferenceDataClientListResult contains the result from method SolutionsReferenceDataClient.List.
-type SolutionsReferenceDataClientListResult struct {
 	SolutionsReferenceDataList
 }
 
 // SubAssessmentsClientGetResponse contains the response from method SubAssessmentsClient.Get.
 type SubAssessmentsClientGetResponse struct {
-	SubAssessmentsClientGetResult
-	// RawResponse contains the underlying HTTP response.
-	RawResponse *http.Response
-}
-
-// SubAssessmentsClientGetResult contains the result from method SubAssessmentsClient.Get.
-type SubAssessmentsClientGetResult struct {
 	SubAssessment
 }
 
 // SubAssessmentsClientListAllResponse contains the response from method SubAssessmentsClient.ListAll.
 type SubAssessmentsClientListAllResponse struct {
-	SubAssessmentsClientListAllResult
-	// RawResponse contains the underlying HTTP response.
-	RawResponse *http.Response
-}
-
-// SubAssessmentsClientListAllResult contains the result from method SubAssessmentsClient.ListAll.
-type SubAssessmentsClientListAllResult struct {
 	SubAssessmentList
 }
 
 // SubAssessmentsClientListResponse contains the response from method SubAssessmentsClient.List.
 type SubAssessmentsClientListResponse struct {
-	SubAssessmentsClientListResult
-	// RawResponse contains the underlying HTTP response.
-	RawResponse *http.Response
-}
-
-// SubAssessmentsClientListResult contains the result from method SubAssessmentsClient.List.
-type SubAssessmentsClientListResult struct {
 	SubAssessmentList
 }
 
 // TasksClientGetResourceGroupLevelTaskResponse contains the response from method TasksClient.GetResourceGroupLevelTask.
 type TasksClientGetResourceGroupLevelTaskResponse struct {
-	TasksClientGetResourceGroupLevelTaskResult
-	// RawResponse contains the underlying HTTP response.
-	RawResponse *http.Response
-}
-
-// TasksClientGetResourceGroupLevelTaskResult contains the result from method TasksClient.GetResourceGroupLevelTask.
-type TasksClientGetResourceGroupLevelTaskResult struct {
 	Task
 }
 
 // TasksClientGetSubscriptionLevelTaskResponse contains the response from method TasksClient.GetSubscriptionLevelTask.
 type TasksClientGetSubscriptionLevelTaskResponse struct {
-	TasksClientGetSubscriptionLevelTaskResult
-	// RawResponse contains the underlying HTTP response.
-	RawResponse *http.Response
-}
-
-// TasksClientGetSubscriptionLevelTaskResult contains the result from method TasksClient.GetSubscriptionLevelTask.
-type TasksClientGetSubscriptionLevelTaskResult struct {
 	Task
 }
 
 // TasksClientListByHomeRegionResponse contains the response from method TasksClient.ListByHomeRegion.
 type TasksClientListByHomeRegionResponse struct {
-	TasksClientListByHomeRegionResult
-	// RawResponse contains the underlying HTTP response.
-	RawResponse *http.Response
-}
-
-// TasksClientListByHomeRegionResult contains the result from method TasksClient.ListByHomeRegion.
-type TasksClientListByHomeRegionResult struct {
 	TaskList
 }
 
 // TasksClientListByResourceGroupResponse contains the response from method TasksClient.ListByResourceGroup.
 type TasksClientListByResourceGroupResponse struct {
-	TasksClientListByResourceGroupResult
-	// RawResponse contains the underlying HTTP response.
-	RawResponse *http.Response
-}
-
-// TasksClientListByResourceGroupResult contains the result from method TasksClient.ListByResourceGroup.
-type TasksClientListByResourceGroupResult struct {
 	TaskList
 }
 
 // TasksClientListResponse contains the response from method TasksClient.List.
 type TasksClientListResponse struct {
-	TasksClientListResult
-	// RawResponse contains the underlying HTTP response.
-	RawResponse *http.Response
-}
-
-// TasksClientListResult contains the result from method TasksClient.List.
-type TasksClientListResult struct {
 	TaskList
 }
 
 // TasksClientUpdateResourceGroupLevelTaskStateResponse contains the response from method TasksClient.UpdateResourceGroupLevelTaskState.
 type TasksClientUpdateResourceGroupLevelTaskStateResponse struct {
-	// RawResponse contains the underlying HTTP response.
-	RawResponse *http.Response
+	// placeholder for future response values
 }
 
 // TasksClientUpdateSubscriptionLevelTaskStateResponse contains the response from method TasksClient.UpdateSubscriptionLevelTaskState.
 type TasksClientUpdateSubscriptionLevelTaskStateResponse struct {
-	// RawResponse contains the underlying HTTP response.
-	RawResponse *http.Response
+	// placeholder for future response values
 }
 
 // TopologyClientGetResponse contains the response from method TopologyClient.Get.
 type TopologyClientGetResponse struct {
-	TopologyClientGetResult
-	// RawResponse contains the underlying HTTP response.
-	RawResponse *http.Response
-}
-
-// TopologyClientGetResult contains the result from method TopologyClient.Get.
-type TopologyClientGetResult struct {
 	TopologyResource
 }
 
 // TopologyClientListByHomeRegionResponse contains the response from method TopologyClient.ListByHomeRegion.
 type TopologyClientListByHomeRegionResponse struct {
-	TopologyClientListByHomeRegionResult
-	// RawResponse contains the underlying HTTP response.
-	RawResponse *http.Response
-}
-
-// TopologyClientListByHomeRegionResult contains the result from method TopologyClient.ListByHomeRegion.
-type TopologyClientListByHomeRegionResult struct {
 	TopologyList
 }
 
 // TopologyClientListResponse contains the response from method TopologyClient.List.
 type TopologyClientListResponse struct {
-	TopologyClientListResult
-	// RawResponse contains the underlying HTTP response.
-	RawResponse *http.Response
-}
-
-// TopologyClientListResult contains the result from method TopologyClient.List.
-type TopologyClientListResult struct {
 	TopologyList
 }
 
 // WorkspaceSettingsClientCreateResponse contains the response from method WorkspaceSettingsClient.Create.
 type WorkspaceSettingsClientCreateResponse struct {
-	WorkspaceSettingsClientCreateResult
-	// RawResponse contains the underlying HTTP response.
-	RawResponse *http.Response
-}
-
-// WorkspaceSettingsClientCreateResult contains the result from method WorkspaceSettingsClient.Create.
-type WorkspaceSettingsClientCreateResult struct {
 	WorkspaceSetting
 }
 
 // WorkspaceSettingsClientDeleteResponse contains the response from method WorkspaceSettingsClient.Delete.
 type WorkspaceSettingsClientDeleteResponse struct {
-	// RawResponse contains the underlying HTTP response.
-	RawResponse *http.Response
+	// placeholder for future response values
 }
 
 // WorkspaceSettingsClientGetResponse contains the response from method WorkspaceSettingsClient.Get.
 type WorkspaceSettingsClientGetResponse struct {
-	WorkspaceSettingsClientGetResult
-	// RawResponse contains the underlying HTTP response.
-	RawResponse *http.Response
-}
-
-// WorkspaceSettingsClientGetResult contains the result from method WorkspaceSettingsClient.Get.
-type WorkspaceSettingsClientGetResult struct {
 	WorkspaceSetting
 }
 
 // WorkspaceSettingsClientListResponse contains the response from method WorkspaceSettingsClient.List.
 type WorkspaceSettingsClientListResponse struct {
-	WorkspaceSettingsClientListResult
-	// RawResponse contains the underlying HTTP response.
-	RawResponse *http.Response
-}
-
-// WorkspaceSettingsClientListResult contains the result from method WorkspaceSettingsClient.List.
-type WorkspaceSettingsClientListResult struct {
 	WorkspaceSettingList
 }
 
 // WorkspaceSettingsClientUpdateResponse contains the response from method WorkspaceSettingsClient.Update.
 type WorkspaceSettingsClientUpdateResponse struct {
-	WorkspaceSettingsClientUpdateResult
-	// RawResponse contains the underlying HTTP response.
-	RawResponse *http.Response
-}
-
-// WorkspaceSettingsClientUpdateResult contains the result from method WorkspaceSettingsClient.Update.
-type WorkspaceSettingsClientUpdateResult struct {
 	WorkspaceSetting
 }
