@@ -1,5 +1,5 @@
-//go:build go1.16
-// +build go1.16
+//go:build go1.18
+// +build go1.18
 
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License. See License.txt in the project root for license information.
@@ -14,6 +14,7 @@ import (
 	"github.com/Azure/azure-sdk-for-go/sdk/azcore"
 	"github.com/Azure/azure-sdk-for-go/sdk/azcore/arm"
 	armruntime "github.com/Azure/azure-sdk-for-go/sdk/azcore/arm/runtime"
+	"github.com/Azure/azure-sdk-for-go/sdk/azcore/cloud"
 	"github.com/Azure/azure-sdk-for-go/sdk/azcore/policy"
 	"github.com/Azure/azure-sdk-for-go/sdk/azcore/runtime"
 	"net/http"
@@ -33,36 +34,56 @@ type SecureScoreControlsClient struct {
 // subscriptionID - Azure subscription ID
 // credential - used to authorize requests. Usually a credential from azidentity.
 // options - pass nil to accept the default values.
-func NewSecureScoreControlsClient(subscriptionID string, credential azcore.TokenCredential, options *arm.ClientOptions) *SecureScoreControlsClient {
+func NewSecureScoreControlsClient(subscriptionID string, credential azcore.TokenCredential, options *arm.ClientOptions) (*SecureScoreControlsClient, error) {
 	if options == nil {
 		options = &arm.ClientOptions{}
 	}
-	ep := options.Endpoint
-	if len(ep) == 0 {
-		ep = arm.AzurePublicCloud
+	ep := cloud.AzurePublicCloud.Services[cloud.ResourceManager].Endpoint
+	if c, ok := options.Cloud.Services[cloud.ResourceManager]; ok {
+		ep = c.Endpoint
+	}
+	pl, err := armruntime.NewPipeline(moduleName, moduleVersion, credential, runtime.PipelineOptions{}, options)
+	if err != nil {
+		return nil, err
 	}
 	client := &SecureScoreControlsClient{
 		subscriptionID: subscriptionID,
-		host:           string(ep),
-		pl:             armruntime.NewPipeline(moduleName, moduleVersion, credential, runtime.PipelineOptions{}, options),
+		host:           ep,
+		pl:             pl,
 	}
-	return client
+	return client, nil
 }
 
 // List - Get all security controls within a scope
 // If the operation fails it returns an *azcore.ResponseError type.
 // options - SecureScoreControlsClientListOptions contains the optional parameters for the SecureScoreControlsClient.List
 // method.
-func (client *SecureScoreControlsClient) List(options *SecureScoreControlsClientListOptions) *SecureScoreControlsClientListPager {
-	return &SecureScoreControlsClientListPager{
-		client: client,
-		requester: func(ctx context.Context) (*policy.Request, error) {
-			return client.listCreateRequest(ctx, options)
+func (client *SecureScoreControlsClient) List(options *SecureScoreControlsClientListOptions) *runtime.Pager[SecureScoreControlsClientListResponse] {
+	return runtime.NewPager(runtime.PageProcessor[SecureScoreControlsClientListResponse]{
+		More: func(page SecureScoreControlsClientListResponse) bool {
+			return page.NextLink != nil && len(*page.NextLink) > 0
 		},
-		advancer: func(ctx context.Context, resp SecureScoreControlsClientListResponse) (*policy.Request, error) {
-			return runtime.NewRequest(ctx, http.MethodGet, *resp.SecureScoreControlList.NextLink)
+		Fetcher: func(ctx context.Context, page *SecureScoreControlsClientListResponse) (SecureScoreControlsClientListResponse, error) {
+			var req *policy.Request
+			var err error
+			if page == nil {
+				req, err = client.listCreateRequest(ctx, options)
+			} else {
+				req, err = runtime.NewRequest(ctx, http.MethodGet, *page.NextLink)
+			}
+			if err != nil {
+				return SecureScoreControlsClientListResponse{}, err
+			}
+			resp, err := client.pl.Do(req)
+			if err != nil {
+				return SecureScoreControlsClientListResponse{}, err
+			}
+			if !runtime.HasStatusCode(resp, http.StatusOK) {
+				return SecureScoreControlsClientListResponse{}, runtime.NewResponseError(resp)
+			}
+			return client.listHandleResponse(resp)
 		},
-	}
+	})
 }
 
 // listCreateRequest creates the List request.
@@ -88,7 +109,7 @@ func (client *SecureScoreControlsClient) listCreateRequest(ctx context.Context, 
 
 // listHandleResponse handles the List response.
 func (client *SecureScoreControlsClient) listHandleResponse(resp *http.Response) (SecureScoreControlsClientListResponse, error) {
-	result := SecureScoreControlsClientListResponse{RawResponse: resp}
+	result := SecureScoreControlsClientListResponse{}
 	if err := runtime.UnmarshalAsJSON(resp, &result.SecureScoreControlList); err != nil {
 		return SecureScoreControlsClientListResponse{}, err
 	}
@@ -100,16 +121,32 @@ func (client *SecureScoreControlsClient) listHandleResponse(resp *http.Response)
 // secureScoreName - The initiative name. For the ASC Default initiative, use 'ascScore' as in the sample request below.
 // options - SecureScoreControlsClientListBySecureScoreOptions contains the optional parameters for the SecureScoreControlsClient.ListBySecureScore
 // method.
-func (client *SecureScoreControlsClient) ListBySecureScore(secureScoreName string, options *SecureScoreControlsClientListBySecureScoreOptions) *SecureScoreControlsClientListBySecureScorePager {
-	return &SecureScoreControlsClientListBySecureScorePager{
-		client: client,
-		requester: func(ctx context.Context) (*policy.Request, error) {
-			return client.listBySecureScoreCreateRequest(ctx, secureScoreName, options)
+func (client *SecureScoreControlsClient) ListBySecureScore(secureScoreName string, options *SecureScoreControlsClientListBySecureScoreOptions) *runtime.Pager[SecureScoreControlsClientListBySecureScoreResponse] {
+	return runtime.NewPager(runtime.PageProcessor[SecureScoreControlsClientListBySecureScoreResponse]{
+		More: func(page SecureScoreControlsClientListBySecureScoreResponse) bool {
+			return page.NextLink != nil && len(*page.NextLink) > 0
 		},
-		advancer: func(ctx context.Context, resp SecureScoreControlsClientListBySecureScoreResponse) (*policy.Request, error) {
-			return runtime.NewRequest(ctx, http.MethodGet, *resp.SecureScoreControlList.NextLink)
+		Fetcher: func(ctx context.Context, page *SecureScoreControlsClientListBySecureScoreResponse) (SecureScoreControlsClientListBySecureScoreResponse, error) {
+			var req *policy.Request
+			var err error
+			if page == nil {
+				req, err = client.listBySecureScoreCreateRequest(ctx, secureScoreName, options)
+			} else {
+				req, err = runtime.NewRequest(ctx, http.MethodGet, *page.NextLink)
+			}
+			if err != nil {
+				return SecureScoreControlsClientListBySecureScoreResponse{}, err
+			}
+			resp, err := client.pl.Do(req)
+			if err != nil {
+				return SecureScoreControlsClientListBySecureScoreResponse{}, err
+			}
+			if !runtime.HasStatusCode(resp, http.StatusOK) {
+				return SecureScoreControlsClientListBySecureScoreResponse{}, runtime.NewResponseError(resp)
+			}
+			return client.listBySecureScoreHandleResponse(resp)
 		},
-	}
+	})
 }
 
 // listBySecureScoreCreateRequest creates the ListBySecureScore request.
@@ -139,7 +176,7 @@ func (client *SecureScoreControlsClient) listBySecureScoreCreateRequest(ctx cont
 
 // listBySecureScoreHandleResponse handles the ListBySecureScore response.
 func (client *SecureScoreControlsClient) listBySecureScoreHandleResponse(resp *http.Response) (SecureScoreControlsClientListBySecureScoreResponse, error) {
-	result := SecureScoreControlsClientListBySecureScoreResponse{RawResponse: resp}
+	result := SecureScoreControlsClientListBySecureScoreResponse{}
 	if err := runtime.UnmarshalAsJSON(resp, &result.SecureScoreControlList); err != nil {
 		return SecureScoreControlsClientListBySecureScoreResponse{}, err
 	}
