@@ -1,5 +1,5 @@
-//go:build go1.16
-// +build go1.16
+//go:build go1.18
+// +build go1.18
 
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License. See License.txt in the project root for license information.
@@ -14,6 +14,7 @@ import (
 	"github.com/Azure/azure-sdk-for-go/sdk/azcore"
 	"github.com/Azure/azure-sdk-for-go/sdk/azcore/arm"
 	armruntime "github.com/Azure/azure-sdk-for-go/sdk/azcore/arm/runtime"
+	"github.com/Azure/azure-sdk-for-go/sdk/azcore/cloud"
 	"github.com/Azure/azure-sdk-for-go/sdk/azcore/policy"
 	"github.com/Azure/azure-sdk-for-go/sdk/azcore/runtime"
 	"net/http"
@@ -33,20 +34,24 @@ type ServerVulnerabilityAssessmentClient struct {
 // subscriptionID - Azure subscription ID
 // credential - used to authorize requests. Usually a credential from azidentity.
 // options - pass nil to accept the default values.
-func NewServerVulnerabilityAssessmentClient(subscriptionID string, credential azcore.TokenCredential, options *arm.ClientOptions) *ServerVulnerabilityAssessmentClient {
+func NewServerVulnerabilityAssessmentClient(subscriptionID string, credential azcore.TokenCredential, options *arm.ClientOptions) (*ServerVulnerabilityAssessmentClient, error) {
 	if options == nil {
 		options = &arm.ClientOptions{}
 	}
-	ep := options.Endpoint
-	if len(ep) == 0 {
-		ep = arm.AzurePublicCloud
+	ep := cloud.AzurePublicCloud.Services[cloud.ResourceManager].Endpoint
+	if c, ok := options.Cloud.Services[cloud.ResourceManager]; ok {
+		ep = c.Endpoint
+	}
+	pl, err := armruntime.NewPipeline(moduleName, moduleVersion, credential, runtime.PipelineOptions{}, options)
+	if err != nil {
+		return nil, err
 	}
 	client := &ServerVulnerabilityAssessmentClient{
 		subscriptionID: subscriptionID,
-		host:           string(ep),
-		pl:             armruntime.NewPipeline(moduleName, moduleVersion, credential, runtime.PipelineOptions{}, options),
+		host:           ep,
+		pl:             pl,
 	}
-	return client
+	return client, nil
 }
 
 // CreateOrUpdate - Creating a server vulnerability assessment on a resource, which will onboard a resource for having a vulnerability
@@ -110,7 +115,7 @@ func (client *ServerVulnerabilityAssessmentClient) createOrUpdateCreateRequest(c
 
 // createOrUpdateHandleResponse handles the CreateOrUpdate response.
 func (client *ServerVulnerabilityAssessmentClient) createOrUpdateHandleResponse(resp *http.Response) (ServerVulnerabilityAssessmentClientCreateOrUpdateResponse, error) {
-	result := ServerVulnerabilityAssessmentClientCreateOrUpdateResponse{RawResponse: resp}
+	result := ServerVulnerabilityAssessmentClientCreateOrUpdateResponse{}
 	if err := runtime.UnmarshalAsJSON(resp, &result.ServerVulnerabilityAssessment); err != nil {
 		return ServerVulnerabilityAssessmentClientCreateOrUpdateResponse{}, err
 	}
@@ -125,22 +130,18 @@ func (client *ServerVulnerabilityAssessmentClient) createOrUpdateHandleResponse(
 // resourceName - Name of the resource.
 // options - ServerVulnerabilityAssessmentClientBeginDeleteOptions contains the optional parameters for the ServerVulnerabilityAssessmentClient.BeginDelete
 // method.
-func (client *ServerVulnerabilityAssessmentClient) BeginDelete(ctx context.Context, resourceGroupName string, resourceNamespace string, resourceType string, resourceName string, options *ServerVulnerabilityAssessmentClientBeginDeleteOptions) (ServerVulnerabilityAssessmentClientDeletePollerResponse, error) {
-	resp, err := client.deleteOperation(ctx, resourceGroupName, resourceNamespace, resourceType, resourceName, options)
-	if err != nil {
-		return ServerVulnerabilityAssessmentClientDeletePollerResponse{}, err
+func (client *ServerVulnerabilityAssessmentClient) BeginDelete(ctx context.Context, resourceGroupName string, resourceNamespace string, resourceType string, resourceName string, options *ServerVulnerabilityAssessmentClientBeginDeleteOptions) (*armruntime.Poller[ServerVulnerabilityAssessmentClientDeleteResponse], error) {
+	if options == nil || options.ResumeToken == "" {
+		resp, err := client.deleteOperation(ctx, resourceGroupName, resourceNamespace, resourceType, resourceName, options)
+		if err != nil {
+			return nil, err
+		}
+		return armruntime.NewPoller(resp, client.pl, &armruntime.NewPollerOptions[ServerVulnerabilityAssessmentClientDeleteResponse]{
+			FinalStateVia: armruntime.FinalStateViaLocation,
+		})
+	} else {
+		return armruntime.NewPollerFromResumeToken[ServerVulnerabilityAssessmentClientDeleteResponse](options.ResumeToken, client.pl, nil)
 	}
-	result := ServerVulnerabilityAssessmentClientDeletePollerResponse{
-		RawResponse: resp,
-	}
-	pt, err := armruntime.NewPoller("ServerVulnerabilityAssessmentClient.Delete", "location", resp, client.pl)
-	if err != nil {
-		return ServerVulnerabilityAssessmentClientDeletePollerResponse{}, err
-	}
-	result.Poller = &ServerVulnerabilityAssessmentClientDeletePoller{
-		pt: pt,
-	}
-	return result, nil
 }
 
 // Delete - Removing server vulnerability assessment from a resource.
@@ -255,7 +256,7 @@ func (client *ServerVulnerabilityAssessmentClient) getCreateRequest(ctx context.
 
 // getHandleResponse handles the Get response.
 func (client *ServerVulnerabilityAssessmentClient) getHandleResponse(resp *http.Response) (ServerVulnerabilityAssessmentClientGetResponse, error) {
-	result := ServerVulnerabilityAssessmentClientGetResponse{RawResponse: resp}
+	result := ServerVulnerabilityAssessmentClientGetResponse{}
 	if err := runtime.UnmarshalAsJSON(resp, &result.ServerVulnerabilityAssessment); err != nil {
 		return ServerVulnerabilityAssessmentClientGetResponse{}, err
 	}
@@ -321,7 +322,7 @@ func (client *ServerVulnerabilityAssessmentClient) listByExtendedResourceCreateR
 
 // listByExtendedResourceHandleResponse handles the ListByExtendedResource response.
 func (client *ServerVulnerabilityAssessmentClient) listByExtendedResourceHandleResponse(resp *http.Response) (ServerVulnerabilityAssessmentClientListByExtendedResourceResponse, error) {
-	result := ServerVulnerabilityAssessmentClientListByExtendedResourceResponse{RawResponse: resp}
+	result := ServerVulnerabilityAssessmentClientListByExtendedResourceResponse{}
 	if err := runtime.UnmarshalAsJSON(resp, &result.ServerVulnerabilityAssessmentsList); err != nil {
 		return ServerVulnerabilityAssessmentClientListByExtendedResourceResponse{}, err
 	}
