@@ -1,5 +1,5 @@
-//go:build go1.16
-// +build go1.16
+//go:build go1.18
+// +build go1.18
 
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License. See License.txt in the project root for license information.
@@ -14,6 +14,7 @@ import (
 	"github.com/Azure/azure-sdk-for-go/sdk/azcore"
 	"github.com/Azure/azure-sdk-for-go/sdk/azcore/arm"
 	armruntime "github.com/Azure/azure-sdk-for-go/sdk/azcore/arm/runtime"
+	"github.com/Azure/azure-sdk-for-go/sdk/azcore/cloud"
 	"github.com/Azure/azure-sdk-for-go/sdk/azcore/policy"
 	"github.com/Azure/azure-sdk-for-go/sdk/azcore/runtime"
 	"net/http"
@@ -31,19 +32,23 @@ type SQLVulnerabilityAssessmentBaselineRulesClient struct {
 // NewSQLVulnerabilityAssessmentBaselineRulesClient creates a new instance of SQLVulnerabilityAssessmentBaselineRulesClient with the specified values.
 // credential - used to authorize requests. Usually a credential from azidentity.
 // options - pass nil to accept the default values.
-func NewSQLVulnerabilityAssessmentBaselineRulesClient(credential azcore.TokenCredential, options *arm.ClientOptions) *SQLVulnerabilityAssessmentBaselineRulesClient {
+func NewSQLVulnerabilityAssessmentBaselineRulesClient(credential azcore.TokenCredential, options *arm.ClientOptions) (*SQLVulnerabilityAssessmentBaselineRulesClient, error) {
 	if options == nil {
 		options = &arm.ClientOptions{}
 	}
-	ep := options.Endpoint
-	if len(ep) == 0 {
-		ep = arm.AzurePublicCloud
+	ep := cloud.AzurePublicCloud.Services[cloud.ResourceManager].Endpoint
+	if c, ok := options.Cloud.Services[cloud.ResourceManager]; ok {
+		ep = c.Endpoint
+	}
+	pl, err := armruntime.NewPipeline(moduleName, moduleVersion, credential, runtime.PipelineOptions{}, options)
+	if err != nil {
+		return nil, err
 	}
 	client := &SQLVulnerabilityAssessmentBaselineRulesClient{
-		host: string(ep),
-		pl:   armruntime.NewPipeline(moduleName, moduleVersion, credential, runtime.PipelineOptions{}, options),
+		host: ep,
+		pl:   pl,
 	}
-	return client
+	return client, nil
 }
 
 // Add - Add a list of baseline rules. Will overwrite any previously existing results (for all rules).
@@ -88,7 +93,7 @@ func (client *SQLVulnerabilityAssessmentBaselineRulesClient) addCreateRequest(ct
 
 // addHandleResponse handles the Add response.
 func (client *SQLVulnerabilityAssessmentBaselineRulesClient) addHandleResponse(resp *http.Response) (SQLVulnerabilityAssessmentBaselineRulesClientAddResponse, error) {
-	result := SQLVulnerabilityAssessmentBaselineRulesClientAddResponse{RawResponse: resp}
+	result := SQLVulnerabilityAssessmentBaselineRulesClientAddResponse{}
 	if err := runtime.UnmarshalAsJSON(resp, &result.RulesResults); err != nil {
 		return SQLVulnerabilityAssessmentBaselineRulesClientAddResponse{}, err
 	}
@@ -142,7 +147,7 @@ func (client *SQLVulnerabilityAssessmentBaselineRulesClient) createOrUpdateCreat
 
 // createOrUpdateHandleResponse handles the CreateOrUpdate response.
 func (client *SQLVulnerabilityAssessmentBaselineRulesClient) createOrUpdateHandleResponse(resp *http.Response) (SQLVulnerabilityAssessmentBaselineRulesClientCreateOrUpdateResponse, error) {
-	result := SQLVulnerabilityAssessmentBaselineRulesClientCreateOrUpdateResponse{RawResponse: resp}
+	result := SQLVulnerabilityAssessmentBaselineRulesClientCreateOrUpdateResponse{}
 	if err := runtime.UnmarshalAsJSON(resp, &result.RuleResults); err != nil {
 		return SQLVulnerabilityAssessmentBaselineRulesClientCreateOrUpdateResponse{}, err
 	}
@@ -168,7 +173,7 @@ func (client *SQLVulnerabilityAssessmentBaselineRulesClient) Delete(ctx context.
 	if !runtime.HasStatusCode(resp, http.StatusOK, http.StatusNoContent) {
 		return SQLVulnerabilityAssessmentBaselineRulesClientDeleteResponse{}, runtime.NewResponseError(resp)
 	}
-	return SQLVulnerabilityAssessmentBaselineRulesClientDeleteResponse{RawResponse: resp}, nil
+	return SQLVulnerabilityAssessmentBaselineRulesClientDeleteResponse{}, nil
 }
 
 // deleteCreateRequest creates the Delete request.
@@ -235,7 +240,7 @@ func (client *SQLVulnerabilityAssessmentBaselineRulesClient) getCreateRequest(ct
 
 // getHandleResponse handles the Get response.
 func (client *SQLVulnerabilityAssessmentBaselineRulesClient) getHandleResponse(resp *http.Response) (SQLVulnerabilityAssessmentBaselineRulesClientGetResponse, error) {
-	result := SQLVulnerabilityAssessmentBaselineRulesClientGetResponse{RawResponse: resp}
+	result := SQLVulnerabilityAssessmentBaselineRulesClientGetResponse{}
 	if err := runtime.UnmarshalAsJSON(resp, &result.RuleResults); err != nil {
 		return SQLVulnerabilityAssessmentBaselineRulesClientGetResponse{}, err
 	}
@@ -281,7 +286,7 @@ func (client *SQLVulnerabilityAssessmentBaselineRulesClient) listCreateRequest(c
 
 // listHandleResponse handles the List response.
 func (client *SQLVulnerabilityAssessmentBaselineRulesClient) listHandleResponse(resp *http.Response) (SQLVulnerabilityAssessmentBaselineRulesClientListResponse, error) {
-	result := SQLVulnerabilityAssessmentBaselineRulesClientListResponse{RawResponse: resp}
+	result := SQLVulnerabilityAssessmentBaselineRulesClientListResponse{}
 	if err := runtime.UnmarshalAsJSON(resp, &result.RulesResults); err != nil {
 		return SQLVulnerabilityAssessmentBaselineRulesClientListResponse{}, err
 	}
