@@ -1,5 +1,5 @@
-//go:build go1.16
-// +build go1.16
+//go:build go1.18
+// +build go1.18
 
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License. See License.txt in the project root for license information.
@@ -14,6 +14,7 @@ import (
 	"github.com/Azure/azure-sdk-for-go/sdk/azcore"
 	"github.com/Azure/azure-sdk-for-go/sdk/azcore/arm"
 	armruntime "github.com/Azure/azure-sdk-for-go/sdk/azcore/arm/runtime"
+	"github.com/Azure/azure-sdk-for-go/sdk/azcore/cloud"
 	"github.com/Azure/azure-sdk-for-go/sdk/azcore/policy"
 	"github.com/Azure/azure-sdk-for-go/sdk/azcore/runtime"
 	"net/http"
@@ -31,19 +32,23 @@ type SQLVulnerabilityAssessmentScanResultsClient struct {
 // NewSQLVulnerabilityAssessmentScanResultsClient creates a new instance of SQLVulnerabilityAssessmentScanResultsClient with the specified values.
 // credential - used to authorize requests. Usually a credential from azidentity.
 // options - pass nil to accept the default values.
-func NewSQLVulnerabilityAssessmentScanResultsClient(credential azcore.TokenCredential, options *arm.ClientOptions) *SQLVulnerabilityAssessmentScanResultsClient {
+func NewSQLVulnerabilityAssessmentScanResultsClient(credential azcore.TokenCredential, options *arm.ClientOptions) (*SQLVulnerabilityAssessmentScanResultsClient, error) {
 	if options == nil {
 		options = &arm.ClientOptions{}
 	}
-	ep := options.Endpoint
-	if len(ep) == 0 {
-		ep = arm.AzurePublicCloud
+	ep := cloud.AzurePublicCloud.Services[cloud.ResourceManager].Endpoint
+	if c, ok := options.Cloud.Services[cloud.ResourceManager]; ok {
+		ep = c.Endpoint
+	}
+	pl, err := armruntime.NewPipeline(moduleName, moduleVersion, credential, runtime.PipelineOptions{}, options)
+	if err != nil {
+		return nil, err
 	}
 	client := &SQLVulnerabilityAssessmentScanResultsClient{
-		host: string(ep),
-		pl:   armruntime.NewPipeline(moduleName, moduleVersion, credential, runtime.PipelineOptions{}, options),
+		host: ep,
+		pl:   pl,
 	}
-	return client
+	return client, nil
 }
 
 // Get - Gets the scan results of a single rule in a scan record.
@@ -95,7 +100,7 @@ func (client *SQLVulnerabilityAssessmentScanResultsClient) getCreateRequest(ctx 
 
 // getHandleResponse handles the Get response.
 func (client *SQLVulnerabilityAssessmentScanResultsClient) getHandleResponse(resp *http.Response) (SQLVulnerabilityAssessmentScanResultsClientGetResponse, error) {
-	result := SQLVulnerabilityAssessmentScanResultsClientGetResponse{RawResponse: resp}
+	result := SQLVulnerabilityAssessmentScanResultsClientGetResponse{}
 	if err := runtime.UnmarshalAsJSON(resp, &result.ScanResult); err != nil {
 		return SQLVulnerabilityAssessmentScanResultsClientGetResponse{}, err
 	}
@@ -146,7 +151,7 @@ func (client *SQLVulnerabilityAssessmentScanResultsClient) listCreateRequest(ctx
 
 // listHandleResponse handles the List response.
 func (client *SQLVulnerabilityAssessmentScanResultsClient) listHandleResponse(resp *http.Response) (SQLVulnerabilityAssessmentScanResultsClientListResponse, error) {
-	result := SQLVulnerabilityAssessmentScanResultsClientListResponse{RawResponse: resp}
+	result := SQLVulnerabilityAssessmentScanResultsClientListResponse{}
 	if err := runtime.UnmarshalAsJSON(resp, &result.ScanResults); err != nil {
 		return SQLVulnerabilityAssessmentScanResultsClientListResponse{}, err
 	}
