@@ -1,5 +1,5 @@
-//go:build go1.16
-// +build go1.16
+//go:build go1.18
+// +build go1.18
 
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License. See License.txt in the project root for license information.
@@ -14,6 +14,7 @@ import (
 	"github.com/Azure/azure-sdk-for-go/sdk/azcore"
 	"github.com/Azure/azure-sdk-for-go/sdk/azcore/arm"
 	armruntime "github.com/Azure/azure-sdk-for-go/sdk/azcore/arm/runtime"
+	"github.com/Azure/azure-sdk-for-go/sdk/azcore/cloud"
 	"github.com/Azure/azure-sdk-for-go/sdk/azcore/policy"
 	"github.com/Azure/azure-sdk-for-go/sdk/azcore/runtime"
 	"net/http"
@@ -31,19 +32,23 @@ type SQLVulnerabilityAssessmentScansClient struct {
 // NewSQLVulnerabilityAssessmentScansClient creates a new instance of SQLVulnerabilityAssessmentScansClient with the specified values.
 // credential - used to authorize requests. Usually a credential from azidentity.
 // options - pass nil to accept the default values.
-func NewSQLVulnerabilityAssessmentScansClient(credential azcore.TokenCredential, options *arm.ClientOptions) *SQLVulnerabilityAssessmentScansClient {
+func NewSQLVulnerabilityAssessmentScansClient(credential azcore.TokenCredential, options *arm.ClientOptions) (*SQLVulnerabilityAssessmentScansClient, error) {
 	if options == nil {
 		options = &arm.ClientOptions{}
 	}
-	ep := options.Endpoint
-	if len(ep) == 0 {
-		ep = arm.AzurePublicCloud
+	ep := cloud.AzurePublicCloud.Services[cloud.ResourceManager].Endpoint
+	if c, ok := options.Cloud.Services[cloud.ResourceManager]; ok {
+		ep = c.Endpoint
+	}
+	pl, err := armruntime.NewPipeline(moduleName, moduleVersion, credential, runtime.PipelineOptions{}, options)
+	if err != nil {
+		return nil, err
 	}
 	client := &SQLVulnerabilityAssessmentScansClient{
-		host: string(ep),
-		pl:   armruntime.NewPipeline(moduleName, moduleVersion, credential, runtime.PipelineOptions{}, options),
+		host: ep,
+		pl:   pl,
 	}
-	return client
+	return client, nil
 }
 
 // Get - Gets the scan details of a single scan record.
@@ -90,7 +95,7 @@ func (client *SQLVulnerabilityAssessmentScansClient) getCreateRequest(ctx contex
 
 // getHandleResponse handles the Get response.
 func (client *SQLVulnerabilityAssessmentScansClient) getHandleResponse(resp *http.Response) (SQLVulnerabilityAssessmentScansClientGetResponse, error) {
-	result := SQLVulnerabilityAssessmentScansClientGetResponse{RawResponse: resp}
+	result := SQLVulnerabilityAssessmentScansClientGetResponse{}
 	if err := runtime.UnmarshalAsJSON(resp, &result.Scan); err != nil {
 		return SQLVulnerabilityAssessmentScansClientGetResponse{}, err
 	}
@@ -136,7 +141,7 @@ func (client *SQLVulnerabilityAssessmentScansClient) listCreateRequest(ctx conte
 
 // listHandleResponse handles the List response.
 func (client *SQLVulnerabilityAssessmentScansClient) listHandleResponse(resp *http.Response) (SQLVulnerabilityAssessmentScansClientListResponse, error) {
-	result := SQLVulnerabilityAssessmentScansClientListResponse{RawResponse: resp}
+	result := SQLVulnerabilityAssessmentScansClientListResponse{}
 	if err := runtime.UnmarshalAsJSON(resp, &result.Scans); err != nil {
 		return SQLVulnerabilityAssessmentScansClientListResponse{}, err
 	}
