@@ -1,5 +1,5 @@
-//go:build go1.16
-// +build go1.16
+//go:build go1.18
+// +build go1.18
 
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License. See License.txt in the project root for license information.
@@ -14,6 +14,7 @@ import (
 	"github.com/Azure/azure-sdk-for-go/sdk/azcore"
 	"github.com/Azure/azure-sdk-for-go/sdk/azcore/arm"
 	armruntime "github.com/Azure/azure-sdk-for-go/sdk/azcore/arm/runtime"
+	"github.com/Azure/azure-sdk-for-go/sdk/azcore/cloud"
 	"github.com/Azure/azure-sdk-for-go/sdk/azcore/policy"
 	"github.com/Azure/azure-sdk-for-go/sdk/azcore/runtime"
 	"net/http"
@@ -33,20 +34,24 @@ type WorkspaceSettingsClient struct {
 // subscriptionID - Azure subscription ID
 // credential - used to authorize requests. Usually a credential from azidentity.
 // options - pass nil to accept the default values.
-func NewWorkspaceSettingsClient(subscriptionID string, credential azcore.TokenCredential, options *arm.ClientOptions) *WorkspaceSettingsClient {
+func NewWorkspaceSettingsClient(subscriptionID string, credential azcore.TokenCredential, options *arm.ClientOptions) (*WorkspaceSettingsClient, error) {
 	if options == nil {
 		options = &arm.ClientOptions{}
 	}
-	ep := options.Endpoint
-	if len(ep) == 0 {
-		ep = arm.AzurePublicCloud
+	ep := cloud.AzurePublicCloud.Services[cloud.ResourceManager].Endpoint
+	if c, ok := options.Cloud.Services[cloud.ResourceManager]; ok {
+		ep = c.Endpoint
+	}
+	pl, err := armruntime.NewPipeline(moduleName, moduleVersion, credential, runtime.PipelineOptions{}, options)
+	if err != nil {
+		return nil, err
 	}
 	client := &WorkspaceSettingsClient{
 		subscriptionID: subscriptionID,
-		host:           string(ep),
-		pl:             armruntime.NewPipeline(moduleName, moduleVersion, credential, runtime.PipelineOptions{}, options),
+		host:           ep,
+		pl:             pl,
 	}
-	return client
+	return client, nil
 }
 
 // Create - creating settings about where we should store your security data and logs
@@ -94,7 +99,7 @@ func (client *WorkspaceSettingsClient) createCreateRequest(ctx context.Context, 
 
 // createHandleResponse handles the Create response.
 func (client *WorkspaceSettingsClient) createHandleResponse(resp *http.Response) (WorkspaceSettingsClientCreateResponse, error) {
-	result := WorkspaceSettingsClientCreateResponse{RawResponse: resp}
+	result := WorkspaceSettingsClientCreateResponse{}
 	if err := runtime.UnmarshalAsJSON(resp, &result.WorkspaceSetting); err != nil {
 		return WorkspaceSettingsClientCreateResponse{}, err
 	}
@@ -118,7 +123,7 @@ func (client *WorkspaceSettingsClient) Delete(ctx context.Context, workspaceSett
 	if !runtime.HasStatusCode(resp, http.StatusNoContent) {
 		return WorkspaceSettingsClientDeleteResponse{}, runtime.NewResponseError(resp)
 	}
-	return WorkspaceSettingsClientDeleteResponse{RawResponse: resp}, nil
+	return WorkspaceSettingsClientDeleteResponse{}, nil
 }
 
 // deleteCreateRequest creates the Delete request.
@@ -187,7 +192,7 @@ func (client *WorkspaceSettingsClient) getCreateRequest(ctx context.Context, wor
 
 // getHandleResponse handles the Get response.
 func (client *WorkspaceSettingsClient) getHandleResponse(resp *http.Response) (WorkspaceSettingsClientGetResponse, error) {
-	result := WorkspaceSettingsClientGetResponse{RawResponse: resp}
+	result := WorkspaceSettingsClientGetResponse{}
 	if err := runtime.UnmarshalAsJSON(resp, &result.WorkspaceSetting); err != nil {
 		return WorkspaceSettingsClientGetResponse{}, err
 	}
@@ -198,16 +203,32 @@ func (client *WorkspaceSettingsClient) getHandleResponse(resp *http.Response) (W
 // configuration was set
 // If the operation fails it returns an *azcore.ResponseError type.
 // options - WorkspaceSettingsClientListOptions contains the optional parameters for the WorkspaceSettingsClient.List method.
-func (client *WorkspaceSettingsClient) List(options *WorkspaceSettingsClientListOptions) *WorkspaceSettingsClientListPager {
-	return &WorkspaceSettingsClientListPager{
-		client: client,
-		requester: func(ctx context.Context) (*policy.Request, error) {
-			return client.listCreateRequest(ctx, options)
+func (client *WorkspaceSettingsClient) List(options *WorkspaceSettingsClientListOptions) *runtime.Pager[WorkspaceSettingsClientListResponse] {
+	return runtime.NewPager(runtime.PageProcessor[WorkspaceSettingsClientListResponse]{
+		More: func(page WorkspaceSettingsClientListResponse) bool {
+			return page.NextLink != nil && len(*page.NextLink) > 0
 		},
-		advancer: func(ctx context.Context, resp WorkspaceSettingsClientListResponse) (*policy.Request, error) {
-			return runtime.NewRequest(ctx, http.MethodGet, *resp.WorkspaceSettingList.NextLink)
+		Fetcher: func(ctx context.Context, page *WorkspaceSettingsClientListResponse) (WorkspaceSettingsClientListResponse, error) {
+			var req *policy.Request
+			var err error
+			if page == nil {
+				req, err = client.listCreateRequest(ctx, options)
+			} else {
+				req, err = runtime.NewRequest(ctx, http.MethodGet, *page.NextLink)
+			}
+			if err != nil {
+				return WorkspaceSettingsClientListResponse{}, err
+			}
+			resp, err := client.pl.Do(req)
+			if err != nil {
+				return WorkspaceSettingsClientListResponse{}, err
+			}
+			if !runtime.HasStatusCode(resp, http.StatusOK) {
+				return WorkspaceSettingsClientListResponse{}, runtime.NewResponseError(resp)
+			}
+			return client.listHandleResponse(resp)
 		},
-	}
+	})
 }
 
 // listCreateRequest creates the List request.
@@ -230,7 +251,7 @@ func (client *WorkspaceSettingsClient) listCreateRequest(ctx context.Context, op
 
 // listHandleResponse handles the List response.
 func (client *WorkspaceSettingsClient) listHandleResponse(resp *http.Response) (WorkspaceSettingsClientListResponse, error) {
-	result := WorkspaceSettingsClientListResponse{RawResponse: resp}
+	result := WorkspaceSettingsClientListResponse{}
 	if err := runtime.UnmarshalAsJSON(resp, &result.WorkspaceSettingList); err != nil {
 		return WorkspaceSettingsClientListResponse{}, err
 	}
@@ -282,7 +303,7 @@ func (client *WorkspaceSettingsClient) updateCreateRequest(ctx context.Context, 
 
 // updateHandleResponse handles the Update response.
 func (client *WorkspaceSettingsClient) updateHandleResponse(resp *http.Response) (WorkspaceSettingsClientUpdateResponse, error) {
-	result := WorkspaceSettingsClientUpdateResponse{RawResponse: resp}
+	result := WorkspaceSettingsClientUpdateResponse{}
 	if err := runtime.UnmarshalAsJSON(resp, &result.WorkspaceSetting); err != nil {
 		return WorkspaceSettingsClientUpdateResponse{}, err
 	}
